+++ runbot_merge/tests/test_basic.py
@@ -10,62 +10,69 @@ from lxml import html
 
 import odoo
 
-from test_utils import re_matches, run_crons, get_partner, _simple_init
+from test_utils import re_matches, get_partner, _simple_init
 
 @pytest.fixture
-def repo(make_repo):
-    return make_repo('repo')
+def repo(project, make_repo):
+    r = make_repo('repo')
+    project.write({'repo_ids': [(0, 0, {'name': r.name})]})
+    return r
 
-def test_trivial_flow(env, repo, page, users):
+def test_trivial_flow(env, repo, page, users, config):
     # create base branch
-    m = repo.make_commit(None, "initial", None, tree={'a': 'some content'})
-    repo.make_ref('heads/master', m)
+    with repo:
+        m = repo.make_commit(None, "initial", None, tree={'a': 'some content'})
+        repo.make_ref('heads/master', m)
 
-    # create PR with 2 commits
-    c0 = repo.make_commit(m, 'replace file contents', None, tree={'a': 'some other content'})
-    c1 = repo.make_commit(c0, 'add file', None, tree={'a': 'some other content', 'b': 'a second file'})
-    pr1 = repo.make_pr("gibberish", "blahblah", target='master', ctid=c1, user='user')
+        # create PR with 2 commits
+        c0 = repo.make_commit(m, 'replace file contents', None, tree={'a': 'some other content'})
+        c1 = repo.make_commit(c0, 'add file', None, tree={'a': 'some other content', 'b': 'a second file'})
+        pr1 = repo.make_pr(title="gibberish", body="blahblah", target='master', head=c1,)
 
     [pr] = env['runbot_merge.pull_requests'].search([
         ('repository.name', '=', repo.name),
         ('number', '=', pr1.number),
     ])
     assert pr.state == 'opened'
-    run_crons(env)
+    env.run_crons()
     assert pr1.labels == {'seen 🙂'}
     # nothing happened
 
-    repo.post_status(c1, 'success', 'legal/cla')
+    with repo:
+        repo.post_status(c1, 'success', 'legal/cla')
     # rewrite status payload in old-style to ensure it does not break
     c = env['runbot_merge.commit'].search([('sha', '=', c1)])
     c.statuses = json.dumps({k: v['state'] for k, v in json.loads(c.statuses).items()})
 
-    repo.post_status(c1, 'success', 'ci/runbot')
+    with repo:
+        repo.post_status(c1, 'success', 'ci/runbot')
 
-    run_crons(env)
+    env.run_crons()
     assert pr.state == 'validated'
 
     assert pr1.labels == {'seen 🙂', 'CI 🤖'}
 
-    pr1.post_comment('hansen r+ rebase-merge', 'reviewer')
+    with repo:
+        pr1.post_comment('hansen r+ rebase-merge', config['role_reviewer']['token'])
     assert pr.state == 'ready'
 
     # can't check labels here as running the cron will stage it
 
-    run_crons(env)
+    env.run_crons()
     assert pr.staging_id
     assert pr1.labels == {'seen 🙂', 'CI 🤖', 'r+ 👌', 'merging 👷'}
 
-    # get head of staging branch
-    staging_head = repo.commit('heads/staging.master')
-    repo.post_status(staging_head.id, 'success', 'ci/runbot', target_url='http://foo.com/pog')
-    repo.post_status(staging_head.id, 'success', 'legal/cla')
-    # the should not block the merge because it's not part of the requirements
-    repo.post_status(staging_head.id, 'failure', 'ci/lint', target_url='http://ignored.com/whocares')
+    with repo:
+        # get head of staging branch
+        staging_head = repo.commit('heads/staging.master')
+        repo.post_status(staging_head.id, 'success', 'ci/runbot', target_url='http://foo.com/pog')
+        repo.post_status(staging_head.id, 'success', 'legal/cla')
+        # the should not block the merge because it's not part of the requirements
+        repo.post_status(staging_head.id, 'failure', 'ci/lint', target_url='http://ignored.com/whocares')
     # need to store this because after the crons have run the staging will
     # have succeeded and been disabled
     st = pr.staging_id
-    run_crons(env)
+    env.run_crons()
 
     assert set(tuple(t) for t in st.statuses) == {
         (repo.name, 'legal/cla', 'success', ''),
@@ -99,46 +106,48 @@ def test_trivial_flow(env, repo, page, users):
                              .format(repo=repo, reviewer=get_partner(env, users['reviewer']))
 
 class TestCommitMessage:
-    def test_commit_simple(self, env, repo, users):
+    def test_commit_simple(self, env, repo, users, config):
         """ verify 'closes ...' is correctly added in the commit message
         """
-        c1 = repo.make_commit(None, 'first!', None, tree={'f': 'm1'})
-        repo.make_ref('heads/master', c1)
-        c2 = repo.make_commit(c1, 'simple commit message', None, tree={'f': 'm2'})
+        with repo:
+            c1 = repo.make_commit(None, 'first!', None, tree={'f': 'm1'})
+            repo.make_ref('heads/master', c1)
+            c2 = repo.make_commit(c1, 'simple commit message', None, tree={'f': 'm2'})
 
-        prx = repo.make_pr('title', 'body', target='master', ctid=c2, user='user')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        prx.post_comment('hansen r+', "reviewer")
+            prx = repo.make_pr(title='title', body='body', target='master', head=c2)
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
+        env.run_crons()
 
-        run_crons(env)
-
-        repo.post_status('heads/staging.master', 'success', 'ci/runbot')
-        repo.post_status('heads/staging.master', 'success', 'legal/cla')
-        run_crons(env)
+        with repo:
+            repo.post_status('heads/staging.master', 'success', 'ci/runbot')
+            repo.post_status('heads/staging.master', 'success', 'legal/cla')
+        env.run_crons()
 
         master = repo.commit('heads/master')
         assert master.message == "simple commit message\n\ncloses {repo.name}#1"\
                                  "\n\nSigned-off-by: {reviewer.formatted_email}"\
                                  .format(repo=repo, reviewer=get_partner(env, users['reviewer']))
 
-    def test_commit_existing(self, env, repo, users):
+    def test_commit_existing(self, env, repo, users, config):
         """ verify do not duplicate 'closes' instruction
         """
-        c1 = repo.make_commit(None, 'first!', None, tree={'f': 'm1'})
-        repo.make_ref('heads/master', c1)
-        c2 = repo.make_commit(c1, 'simple commit message that closes #1', None, tree={'f': 'm2'})
+        with repo:
+            c1 = repo.make_commit(None, 'first!', None, tree={'f': 'm1'})
+            repo.make_ref('heads/master', c1)
+            c2 = repo.make_commit(c1, 'simple commit message that closes #1', None, tree={'f': 'm2'})
 
-        prx = repo.make_pr('title', 'body', target='master', ctid=c2, user='user')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        prx.post_comment('hansen r+', "reviewer")
+            prx = repo.make_pr(title='title', body='body', target='master', head=c2)
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
+        env.run_crons()
 
-        run_crons(env)
-
-        repo.post_status('heads/staging.master', 'success', 'ci/runbot')
-        repo.post_status('heads/staging.master', 'success', 'legal/cla')
-        run_crons(env)
+        with repo:
+            repo.post_status('heads/staging.master', 'success', 'ci/runbot')
+            repo.post_status('heads/staging.master', 'success', 'legal/cla')
+        env.run_crons()
 
         master = repo.commit('heads/master')
         # closes #1 is already present, should not modify message
@@ -146,23 +155,24 @@ class TestCommitMessage:
                                  "\n\nSigned-off-by: {reviewer.formatted_email}"\
                                  .format(reviewer=get_partner(env, users['reviewer']))
 
-    def test_commit_other(self, env, repo, users):
+    def test_commit_other(self, env, repo, users, config):
         """ verify do not duplicate 'closes' instruction
         """
-        c1 = repo.make_commit(None, 'first!', None, tree={'f': 'm1'})
-        repo.make_ref('heads/master', c1)
-        c2 = repo.make_commit(c1, 'simple commit message that closes odoo/enterprise#1', None, tree={'f': 'm2'})
+        with repo:
+            c1 = repo.make_commit(None, 'first!', None, tree={'f': 'm1'})
+            repo.make_ref('heads/master', c1)
+            c2 = repo.make_commit(c1, 'simple commit message that closes odoo/enterprise#1', None, tree={'f': 'm2'})
 
-        prx = repo.make_pr('title', 'body', target='master', ctid=c2, user='user')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        prx.post_comment('hansen r+', "reviewer")
+            prx = repo.make_pr(title='title', body='body', target='master', head=c2)
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
+        env.run_crons()
 
-        run_crons(env)
-
-        repo.post_status('heads/staging.master', 'success', 'ci/runbot')
-        repo.post_status('heads/staging.master', 'success', 'legal/cla')
-        run_crons(env)
+        with repo:
+            repo.post_status('heads/staging.master', 'success', 'ci/runbot')
+            repo.post_status('heads/staging.master', 'success', 'legal/cla')
+        env.run_crons()
 
         master = repo.commit('heads/master')
         # closes on another repositoy, should modify the commit message
@@ -170,23 +180,24 @@ class TestCommitMessage:
                                  "\n\nSigned-off-by: {reviewer.formatted_email}"\
                                  .format(repo=repo, reviewer=get_partner(env, users['reviewer']))
 
-    def test_commit_wrong_number(self, env, repo, users):
+    def test_commit_wrong_number(self, env, repo, users, config):
         """ verify do not match on a wrong number
         """
-        c1 = repo.make_commit(None, 'first!', None, tree={'f': 'm1'})
-        repo.make_ref('heads/master', c1)
-        c2 = repo.make_commit(c1, 'simple commit message that closes #11', None, tree={'f': 'm2'})
+        with repo:
+            c1 = repo.make_commit(None, 'first!', None, tree={'f': 'm1'})
+            repo.make_ref('heads/master', c1)
+            c2 = repo.make_commit(c1, 'simple commit message that closes #11', None, tree={'f': 'm2'})
 
-        prx = repo.make_pr('title', 'body', target='master', ctid=c2, user='user')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        prx.post_comment('hansen r+', "reviewer")
+            prx = repo.make_pr(title='title', body='body', target='master', head=c2)
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
+        env.run_crons()
 
-        run_crons(env)
-
-        repo.post_status('heads/staging.master', 'success', 'ci/runbot')
-        repo.post_status('heads/staging.master', 'success', 'legal/cla')
-        run_crons(env)
+        with repo:
+            repo.post_status('heads/staging.master', 'success', 'ci/runbot')
+            repo.post_status('heads/staging.master', 'success', 'legal/cla')
+        env.run_crons()
 
         master = repo.commit('heads/master')
         # closes on another repositoy, should modify the commit message
@@ -194,55 +205,57 @@ class TestCommitMessage:
                                  "\n\nSigned-off-by: {reviewer.formatted_email}"\
                                  .format(repo=repo, reviewer=get_partner(env, users['reviewer']))
 
-    def test_commit_delegate(self, env, repo, users):
+    def test_commit_delegate(self, env, repo, users, config):
         """ verify 'signed-off-by ...' is correctly added in the commit message for delegated review
         """
-        c1 = repo.make_commit(None, 'first!', None, tree={'f': 'm1'})
-        repo.make_ref('heads/master', c1)
-        c2 = repo.make_commit(c1, 'simple commit message', None, tree={'f': 'm2'})
+        with repo:
+            c1 = repo.make_commit(None, 'first!', None, tree={'f': 'm1'})
+            repo.make_ref('heads/master', c1)
+            c2 = repo.make_commit(c1, 'simple commit message', None, tree={'f': 'm2'})
 
-        prx = repo.make_pr('title', 'body', target='master', ctid=c2, user='user')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        prx.post_comment('hansen delegate=%s' % users['other'], "reviewer")
-        prx.post_comment('hansen r+', user='other')
+            prx = repo.make_pr(title='title', body='body', target='master', head=c2)
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            prx.post_comment('hansen delegate=%s' % users['other'], config["role_reviewer"]["token"])
+            prx.post_comment('hansen r+', config['role_other']['token'])
+        env.run_crons()
 
-        run_crons(env)
-
-        repo.post_status('heads/staging.master', 'success', 'ci/runbot')
-        repo.post_status('heads/staging.master', 'success', 'legal/cla')
-        run_crons(env)
+        with repo:
+            repo.post_status('heads/staging.master', 'success', 'ci/runbot')
+            repo.post_status('heads/staging.master', 'success', 'legal/cla')
+        env.run_crons()
 
         master = repo.commit('heads/master')
         assert master.message == "simple commit message\n\ncloses {repo.name}#1"\
                                  "\n\nSigned-off-by: {reviewer.formatted_email}"\
                                  .format(repo=repo, reviewer=get_partner(env, users['other']))
 
-    def test_commit_coauthored(self, env, repo, users):
+    def test_commit_coauthored(self, env, repo, users, config):
         """ verify 'closes ...' and 'Signed-off-by' are added before co-authored-by tags.
 
         Also checks that all co-authored-by are moved at the end of the
         message
         """
-        c1 = repo.make_commit(None, 'first!', None, tree={'f': 'm1'})
-        repo.make_ref('heads/master', c1)
-        c2 = repo.make_commit(c1, '''simple commit message
+        with repo:
+            c1 = repo.make_commit(None, 'first!', None, tree={'f': 'm1'})
+            repo.make_ref('heads/master', c1)
+            c2 = repo.make_commit(c1, '''simple commit message
 
 
 Co-authored-by: Bob <bob@example.com>
 
 Fixes a thing''', None, tree={'f': 'm2'})
 
-        prx = repo.make_pr('title', 'body', target='master', ctid=c2, user='user')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        prx.post_comment('hansen r+', "reviewer")
+            prx = repo.make_pr(title='title', body='body', target='master', head=c2)
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
+        env.run_crons()
 
-        run_crons(env)
-
-        repo.post_status('heads/staging.master', 'success', 'ci/runbot')
-        repo.post_status('heads/staging.master', 'success', 'legal/cla')
-        run_crons(env)
+        with repo:
+            repo.post_status('heads/staging.master', 'success', 'ci/runbot')
+            repo.post_status('heads/staging.master', 'success', 'legal/cla')
+        env.run_crons()
 
         master = repo.commit('heads/master')
         assert master.message == """simple commit message
@@ -263,11 +276,12 @@ class TestWebhookSecurity:
         """
         project.secret = "a secret"
 
-        m = repo.make_commit(None, "initial", None, tree={'a': 'some content'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, "initial", None, tree={'a': 'some content'})
+            repo.make_ref('heads/master', m)
 
-        c0 = repo.make_commit(m, 'replace file contents', None, tree={'a': 'some other content'})
-        pr0 = repo.make_pr("gibberish", "blahblah", target='master', ctid=c0, user='user')
+            c0 = repo.make_commit(m, 'replace file contents', None, tree={'a': 'some other content'})
+            pr0 = repo.make_pr(title="gibberish", body="blahblah", target='master', head=c0)
 
         assert not env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
@@ -275,14 +289,15 @@ class TestWebhookSecurity:
         ])
 
     def test_wrong_secret(self, env, project, repo):
-        repo.set_secret("wrong secret")
         project.secret = "a secret"
+        with repo:
+            repo.set_secret("wrong secret")
 
-        m = repo.make_commit(None, "initial", None, tree={'a': 'some content'})
-        repo.make_ref('heads/master', m)
+            m = repo.make_commit(None, "initial", None, tree={'a': 'some content'})
+            repo.make_ref('heads/master', m)
 
-        c0 = repo.make_commit(m, 'replace file contents', None, tree={'a': 'some other content'})
-        pr0 = repo.make_pr("gibberish", "blahblah", target='master', ctid=c0, user='user')
+            c0 = repo.make_commit(m, 'replace file contents', None, tree={'a': 'some other content'})
+            pr0 = repo.make_pr(title="gibberish", body="blahblah", target='master', head=c0)
 
         assert not env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
@@ -290,47 +305,50 @@ class TestWebhookSecurity:
         ])
 
     def test_correct_secret(self, env, project, repo):
-        repo.set_secret("a secret")
         project.secret = "a secret"
+        with repo:
+            repo.set_secret("a secret")
 
-        m = repo.make_commit(None, "initial", None, tree={'a': 'some content'})
-        repo.make_ref('heads/master', m)
+            m = repo.make_commit(None, "initial", None, tree={'a': 'some content'})
+            repo.make_ref('heads/master', m)
 
-        c0 = repo.make_commit(m, 'replace file contents', None, tree={'a': 'some other content'})
-        pr0 = repo.make_pr("gibberish", "blahblah", target='master', ctid=c0, user='user')
+            c0 = repo.make_commit(m, 'replace file contents', None, tree={'a': 'some other content'})
+            pr0 = repo.make_pr(title="gibberish", body="blahblah", target='master', head=c0)
 
         assert env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', pr0.number),
         ])
 
-def test_staging_conflict(env, repo):
-    # create base branch
-    m = repo.make_commit(None, 'initial', None, tree={'a': 'some content'})
-    repo.make_ref('heads/master', m)
+def test_staging_conflict(env, repo, config):
+    with repo:
+        # create base branch
+        m = repo.make_commit(None, 'initial', None, tree={'a': 'some content'})
+        repo.make_ref('heads/master', m)
 
-    # create PR
-    c0 = repo.make_commit(m, 'replace file contents', None, tree={'a': 'some other content'})
-    c1 = repo.make_commit(c0, 'add file', None, tree={'a': 'some other content', 'b': 'a second file'})
-    pr1 = repo.make_pr("gibberish", "blahblah", target='master', ctid=c1, user='user')
-    repo.post_status(c1, 'success', 'legal/cla')
-    repo.post_status(c1, 'success', 'ci/runbot')
-    pr1.post_comment("hansen r+ rebase-merge", "reviewer")
-    run_crons(env)
+        # create PR
+        c0 = repo.make_commit(m, 'replace file contents', None, tree={'a': 'some other content'})
+        c1 = repo.make_commit(c0, 'add file', None, tree={'a': 'some other content', 'b': 'a second file'})
+        pr1 = repo.make_pr(title="gibberish", body="blahblah", target='master', head=c1)
+        repo.post_status(c1, 'success', 'legal/cla')
+        repo.post_status(c1, 'success', 'ci/runbot')
+        pr1.post_comment("hansen r+ rebase-merge", config['role_reviewer']['token'])
+    env.run_crons()
     pr1 = env['runbot_merge.pull_requests'].search([
         ('repository.name', '=', repo.name),
         ('number', '=', 1)
     ])
     assert pr1.staging_id
 
-    # create second PR and make ready for staging
-    c2 = repo.make_commit(m, 'other', None, tree={'a': 'some content', 'c': 'ccc'})
-    c3 = repo.make_commit(c2, 'other', None, tree={'a': 'some content', 'c': 'ccc', 'd': 'ddd'})
-    pr2 = repo.make_pr('gibberish', 'blahblah', target='master', ctid=c3, user='user')
-    repo.post_status(c3, 'success', 'legal/cla')
-    repo.post_status(c3, 'success', 'ci/runbot')
-    pr2.post_comment('hansen r+ rebase-merge', "reviewer")
-    run_crons(env)
+    with repo:
+        # create second PR and make ready for staging
+        c2 = repo.make_commit(m, 'other', None, tree={'a': 'some content', 'c': 'ccc'})
+        c3 = repo.make_commit(c2, 'other', None, tree={'a': 'some content', 'c': 'ccc', 'd': 'ddd'})
+        pr2 = repo.make_pr(title='gibberish', body='blahblah', target='master', head=c3)
+        repo.post_status(c3, 'success', 'legal/cla')
+        repo.post_status(c3, 'success', 'ci/runbot')
+        pr2.post_comment('hansen r+ rebase-merge', config['role_reviewer']['token'])
+    env.run_crons()
     p_2 = env['runbot_merge.pull_requests'].search([
         ('repository.name', '=', repo.name),
         ('number', '=', pr2.number)
@@ -339,43 +357,47 @@ def test_staging_conflict(env, repo):
     assert pr2.labels == {'seen 🙂', 'CI 🤖', 'r+ 👌'}
 
     staging_head = repo.commit('heads/staging.master')
-    repo.post_status(staging_head.id, 'success', 'ci/runbot')
-    repo.post_status(staging_head.id, 'success', 'legal/cla')
-    run_crons(env)
+    with repo:
+        repo.post_status(staging_head.id, 'success', 'ci/runbot')
+        repo.post_status(staging_head.id, 'success', 'legal/cla')
+    env.run_crons()
     assert pr1.state == 'merged'
     assert p_2.staging_id
 
     staging_head = repo.commit('heads/staging.master')
-    repo.post_status(staging_head.id, 'success', 'ci/runbot')
-    repo.post_status(staging_head.id, 'success', 'legal/cla')
-    run_crons(env)
+    with repo:
+        repo.post_status(staging_head.id, 'success', 'ci/runbot')
+        repo.post_status(staging_head.id, 'success', 'legal/cla')
+    env.run_crons()
     assert p_2.state == 'merged'
 
-def test_staging_concurrent(env, repo):
+def test_staging_concurrent(env, repo, config):
     """ test staging to different targets, should be picked up together """
-    m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-    repo.make_ref('heads/1.0', m)
-    repo.make_ref('heads/2.0', m)
+    with repo:
+        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+        repo.make_ref('heads/1.0', m)
+        repo.make_ref('heads/2.0', m)
 
     env['runbot_merge.project'].search([]).write({
         'branch_ids': [(0, 0, {'name': '1.0'}), (0, 0, {'name': '2.0'})],
     })
 
-    c10 = repo.make_commit(m, 'AAA', None, tree={'m': 'm', 'a': 'a'})
-    c11 = repo.make_commit(c10, 'BBB', None, tree={'m': 'm', 'a': 'a', 'b': 'b'})
-    pr1 = repo.make_pr('t1', 'b1', target='1.0', ctid=c11, user='user')
-    repo.post_status(pr1.head, 'success', 'ci/runbot')
-    repo.post_status(pr1.head, 'success', 'legal/cla')
-    pr1.post_comment('hansen r+ rebase-merge', "reviewer")
+    with repo:
+        c10 = repo.make_commit(m, 'AAA', None, tree={'m': 'm', 'a': 'a'})
+        c11 = repo.make_commit(c10, 'BBB', None, tree={'m': 'm', 'a': 'a', 'b': 'b'})
+        pr1 = repo.make_pr(title='t1', body='b1', target='1.0', head=c11)
+        repo.post_status(pr1.head, 'success', 'ci/runbot')
+        repo.post_status(pr1.head, 'success', 'legal/cla')
+        pr1.post_comment('hansen r+ rebase-merge', config['role_reviewer']['token'])
 
-    c20 = repo.make_commit(m, 'CCC', None, tree={'m': 'm', 'c': 'c'})
-    c21 = repo.make_commit(c20, 'DDD', None, tree={'m': 'm', 'c': 'c', 'd': 'd'})
-    pr2 = repo.make_pr('t2', 'b2', target='2.0', ctid=c21, user='user')
-    repo.post_status(pr2.head, 'success', 'ci/runbot')
-    repo.post_status(pr2.head, 'success', 'legal/cla')
-    pr2.post_comment('hansen r+ rebase-merge', "reviewer")
+        c20 = repo.make_commit(m, 'CCC', None, tree={'m': 'm', 'c': 'c'})
+        c21 = repo.make_commit(c20, 'DDD', None, tree={'m': 'm', 'c': 'c', 'd': 'd'})
+        pr2 = repo.make_pr(title='t2', body='b2', target='2.0', head=c21)
+        repo.post_status(pr2.head, 'success', 'ci/runbot')
+        repo.post_status(pr2.head, 'success', 'legal/cla')
+        pr2.post_comment('hansen r+ rebase-merge', config['role_reviewer']['token'])
+    env.run_crons()
 
-    run_crons(env)
     pr1 = env['runbot_merge.pull_requests'].search([
         ('repository.name', '=', repo.name),
         ('number', '=', pr1.number)
@@ -387,21 +409,22 @@ def test_staging_concurrent(env, repo):
     ])
     assert pr2.staging_id
 
-def test_staging_merge_fail(env, repo, users):
+def test_staging_merge_fail(env, repo, users, config):
     """ # of staging failure (no CI) before mark & notify?
     """
-    m1 = repo.make_commit(None, 'initial', None, tree={'f': 'm1'})
-    m2 = repo.make_commit(m1, 'second', None, tree={'f': 'm2'})
-    repo.make_ref('heads/master', m2)
+    with repo:
+        m1 = repo.make_commit(None, 'initial', None, tree={'f': 'm1'})
+        m2 = repo.make_commit(m1, 'second', None, tree={'f': 'm2'})
+        repo.make_ref('heads/master', m2)
 
-    c1 = repo.make_commit(m1, 'other second', None, tree={'f': 'c1'})
-    c2 = repo.make_commit(c1, 'third', None, tree={'f': 'c2'})
-    prx = repo.make_pr('title', 'body', target='master', ctid=c2, user='user')
-    repo.post_status(prx.head, 'success', 'ci/runbot')
-    repo.post_status(prx.head, 'success', 'legal/cla')
-    prx.post_comment('hansen r+ rebase-merge', "reviewer")
+        c1 = repo.make_commit(m1, 'other second', None, tree={'f': 'c1'})
+        c2 = repo.make_commit(c1, 'third', None, tree={'f': 'c2'})
+        prx = repo.make_pr(title='title', body='body', target='master', head=c2)
+        repo.post_status(prx.head, 'success', 'ci/runbot')
+        repo.post_status(prx.head, 'success', 'legal/cla')
+        prx.post_comment('hansen r+ rebase-merge', config['role_reviewer']['token'])
+    env.run_crons()
 
-    run_crons(env)
     pr1 = env['runbot_merge.pull_requests'].search([
         ('repository.name', '=', repo.name),
         ('number', '=', prx.number)
@@ -414,20 +437,21 @@ def test_staging_merge_fail(env, repo, users):
         (users['user'], re_matches('^Unable to stage PR')),
     ]
 
-def test_staging_ci_timeout(env, repo, users):
+def test_staging_ci_timeout(env, repo, config):
     """If a staging timeouts (~ delay since staged greater than
     configured)... requeue?
     """
-    m = repo.make_commit(None, 'initial', None, tree={'f': 'm'})
-    repo.make_ref('heads/master', m)
+    with repo:
+        m = repo.make_commit(None, 'initial', None, tree={'f': 'm'})
+        repo.make_ref('heads/master', m)
 
-    c1 = repo.make_commit(m, 'first', None, tree={'f': 'c1'})
-    c2 = repo.make_commit(c1, 'second', None, tree={'f': 'c2'})
-    prx = repo.make_pr('title', 'body', target='master', ctid=c2, user='user')
-    repo.post_status(prx.head, 'success', 'ci/runbot')
-    repo.post_status(prx.head, 'success', 'legal/cla')
-    prx.post_comment('hansen r+ rebase-merge', "reviewer")
-    run_crons(env)
+        c1 = repo.make_commit(m, 'first', None, tree={'f': 'c1'})
+        c2 = repo.make_commit(c1, 'second', None, tree={'f': 'c2'})
+        prx = repo.make_pr(title='title', body='body', target='master', head=c2)
+        repo.post_status(prx.head, 'success', 'ci/runbot')
+        repo.post_status(prx.head, 'success', 'legal/cla')
+        prx.post_comment('hansen r+ rebase-merge', config['role_reviewer']['token'])
+    env.run_crons()
 
     pr1 = env['runbot_merge.pull_requests'].search([
         ('repository.name', '=', repo.name),
@@ -437,49 +461,53 @@ def test_staging_ci_timeout(env, repo, users):
     timeout = env['runbot_merge.project'].search([]).ci_timeout
 
     pr1.staging_id.staged_at = odoo.fields.Datetime.to_string(datetime.datetime.now() - datetime.timedelta(minutes=2*timeout))
-    env['runbot_merge.project']._check_progress()
+    env.run_crons('runbot_merge.merge_cron')
     assert pr1.state == 'error', "timeout should fail the PR"
 
-def test_timeout_bump_on_pending(env, repo):
-    m = repo.make_commit(None, 'initial', None, tree={'f': '0'})
-    repo.make_ref('heads/master', m)
+def test_timeout_bump_on_pending(env, repo, config):
+    with repo:
+        m = repo.make_commit(None, 'initial', None, tree={'f': '0'})
+        repo.make_ref('heads/master', m)
 
-    c = repo.make_commit(m, 'c', None, tree={'f': '1'})
-    prx = repo.make_pr('title', 'body', target='master', ctid=c, user='user')
-    repo.post_status(prx.head, 'success', 'ci/runbot')
-    repo.post_status(prx.head, 'success', 'legal/cla')
-    prx.post_comment('hansen r+', 'reviewer')
-    run_crons(env)
+        c = repo.make_commit(m, 'c', None, tree={'f': '1'})
+        prx = repo.make_pr(title='title', body='body', target='master', head=c)
+        repo.post_status(prx.head, 'success', 'ci/runbot')
+        repo.post_status(prx.head, 'success', 'legal/cla')
+        prx.post_comment('hansen r+', config['role_reviewer']['token'])
+    env.run_crons()
 
     st = env['runbot_merge.stagings'].search([])
     old_timeout = odoo.fields.Datetime.to_string(datetime.datetime.now() - datetime.timedelta(days=15))
     st.timeout_limit = old_timeout
-    repo.post_status(repo.commit('heads/staging.master').id, 'pending', 'ci/runbot')
-    env['runbot_merge.commit']._notify()
+    with repo:
+        repo.post_status(repo.commit('heads/staging.master').id, 'pending', 'ci/runbot')
+    env.run_crons('runbot_merge.process_updated_commits')
     assert st.timeout_limit > old_timeout
 
-def test_staging_ci_failure_single(env, repo, users):
+def test_staging_ci_failure_single(env, repo, users, config):
     """ on failure of single-PR staging, mark & notify failure
     """
-    m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-    repo.make_ref('heads/master', m)
+    with repo:
+        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+        repo.make_ref('heads/master', m)
 
-    c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
-    c2 = repo.make_commit(c1, 'second', None, tree={'m': 'c2'})
-    prx = repo.make_pr('title', 'body', target='master', ctid=c2, user='user')
-    repo.post_status(prx.head, 'success', 'ci/runbot')
-    repo.post_status(prx.head, 'success', 'legal/cla')
-    prx.post_comment('hansen r+ rebase-merge', "reviewer")
-    run_crons(env)
+        c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
+        c2 = repo.make_commit(c1, 'second', None, tree={'m': 'c2'})
+        prx = repo.make_pr(title='title', body='body', target='master', head=c2)
+        repo.post_status(prx.head, 'success', 'ci/runbot')
+        repo.post_status(prx.head, 'success', 'legal/cla')
+        prx.post_comment('hansen r+ rebase-merge', config['role_reviewer']['token'])
+    env.run_crons()
     assert env['runbot_merge.pull_requests'].search([
         ('repository.name', '=', repo.name),
         ('number', '=', prx.number)
     ]).staging_id
 
     staging_head = repo.commit('heads/staging.master')
-    repo.post_status(staging_head.id, 'success', 'legal/cla')
-    repo.post_status(staging_head.id, 'failure', 'ci/runbot') # stable genius
-    run_crons(env)
+    with repo:
+        repo.post_status(staging_head.id, 'success', 'legal/cla')
+        repo.post_status(staging_head.id, 'failure', 'ci/runbot') # stable genius
+    env.run_crons()
     assert env['runbot_merge.pull_requests'].search([
         ('repository.name', '=', repo.name),
         ('number', '=', prx.number)
@@ -491,31 +519,34 @@ def test_staging_ci_failure_single(env, repo, users):
         (users['user'], 'Staging failed: ci/runbot')
     ]
 
-def test_ff_failure(env, repo, users):
+def test_ff_failure(env, repo, config):
     """ target updated while the PR is being staged => redo staging """
-    m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-    repo.make_ref('heads/master', m)
+    with repo:
+        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+        repo.make_ref('heads/master', m)
 
-    c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
-    c2 = repo.make_commit(c1, 'second', None, tree={'m': 'c2'})
-    prx = repo.make_pr('title', 'body', target='master', ctid=c2, user='user')
-    repo.post_status(prx.head, 'success', 'legal/cla')
-    repo.post_status(prx.head, 'success', 'ci/runbot')
-    prx.post_comment('hansen r+ rebase-merge', "reviewer")
-    run_crons(env)
+        c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
+        c2 = repo.make_commit(c1, 'second', None, tree={'m': 'c2'})
+        prx = repo.make_pr(title='title', body='body', target='master', head=c2)
+        repo.post_status(prx.head, 'success', 'legal/cla')
+        repo.post_status(prx.head, 'success', 'ci/runbot')
+        prx.post_comment('hansen r+ rebase-merge', config['role_reviewer']['token'])
+    env.run_crons()
     assert env['runbot_merge.pull_requests'].search([
         ('repository.name', '=', repo.name),
         ('number', '=', prx.number)
     ]).staging_id
 
-    m2 = repo.make_commit('heads/master', 'cockblock', None, tree={'m': 'm', 'm2': 'm2'})
+    with repo:
+        m2 = repo.make_commit('heads/master', 'cockblock', None, tree={'m': 'm', 'm2': 'm2'})
     assert repo.commit('heads/master').id == m2
 
     # report staging success & run cron to merge
     staging = repo.commit('heads/staging.master')
-    repo.post_status(staging.id, 'success', 'legal/cla')
-    repo.post_status(staging.id, 'success', 'ci/runbot')
-    run_crons(env)
+    with repo:
+        repo.post_status(staging.id, 'success', 'legal/cla')
+        repo.post_status(staging.id, 'success', 'ci/runbot')
+    env.run_crons()
 
     assert env['runbot_merge.pull_requests'].search([
         ('repository.name', '=', repo.name),
@@ -524,32 +555,36 @@ def test_ff_failure(env, repo, users):
     assert repo.commit('heads/staging.master').id != staging.id,\
         "PR should be staged to a new commit"
 
-def test_ff_failure_batch(env, repo, users):
-    m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-    repo.make_ref('heads/master', m)
+def test_ff_failure_batch(env, repo, users, config):
+    with repo:
+        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+        repo.make_ref('heads/master', m)
 
-    a1 = repo.make_commit(m, 'a1', None, tree={'m': 'm', 'a': '1'})
-    a2 = repo.make_commit(a1, 'a2', None, tree={'m': 'm', 'a': '2'})
-    A = repo.make_pr('A', None, target='master', ctid=a2, user='user', label='A')
-    repo.post_status(A.head, 'success', 'legal/cla')
-    repo.post_status(A.head, 'success', 'ci/runbot')
-    A.post_comment('hansen r+ rebase-merge', "reviewer")
+        a1 = repo.make_commit(m, 'a1', None, tree={'m': 'm', 'a': '1'})
+        a2 = repo.make_commit(a1, 'a2', None, tree={'m': 'm', 'a': '2'})
+        repo.make_ref('heads/A', a2)
+        A = repo.make_pr(title='A', body=None, target='master', head='A')
+        repo.post_status(A.head, 'success', 'legal/cla')
+        repo.post_status(A.head, 'success', 'ci/runbot')
+        A.post_comment('hansen r+ rebase-merge', config['role_reviewer']['token'])
 
-    b1 = repo.make_commit(m, 'b1', None, tree={'m': 'm', 'b': '1'})
-    b2 = repo.make_commit(b1, 'b2', None, tree={'m': 'm', 'b': '2'})
-    B = repo.make_pr('B', None, target='master', ctid=b2, user='user', label='B')
-    repo.post_status(B.head, 'success', 'legal/cla')
-    repo.post_status(B.head, 'success', 'ci/runbot')
-    B.post_comment('hansen r+ rebase-merge', "reviewer")
+        b1 = repo.make_commit(m, 'b1', None, tree={'m': 'm', 'b': '1'})
+        b2 = repo.make_commit(b1, 'b2', None, tree={'m': 'm', 'b': '2'})
+        repo.make_ref('heads/B', b2)
+        B = repo.make_pr(title='B', body=None, target='master', head='B')
+        repo.post_status(B.head, 'success', 'legal/cla')
+        repo.post_status(B.head, 'success', 'ci/runbot')
+        B.post_comment('hansen r+ rebase-merge', config['role_reviewer']['token'])
 
-    c1 = repo.make_commit(m, 'c1', None, tree={'m': 'm', 'c': '1'})
-    c2 = repo.make_commit(c1, 'c2', None, tree={'m': 'm', 'c': '2'})
-    C = repo.make_pr('C', None, target='master', ctid=c2, user='user', label='C')
-    repo.post_status(C.head, 'success', 'legal/cla')
-    repo.post_status(C.head, 'success', 'ci/runbot')
-    C.post_comment('hansen r+ rebase-merge', "reviewer")
+        c1 = repo.make_commit(m, 'c1', None, tree={'m': 'm', 'c': '1'})
+        c2 = repo.make_commit(c1, 'c2', None, tree={'m': 'm', 'c': '2'})
+        repo.make_ref('heads/C', c2)
+        C = repo.make_pr(title='C', body=None, target='master', head='C')
+        repo.post_status(C.head, 'success', 'legal/cla')
+        repo.post_status(C.head, 'success', 'ci/runbot')
+        C.post_comment('hansen r+ rebase-merge', config['role_reviewer']['token'])
+    env.run_crons()
 
-    run_crons(env)
     messages = [
         c['commit']['message']
         for c in repo.log('heads/staging.master')
@@ -559,21 +594,24 @@ def test_ff_failure_batch(env, repo, users):
     assert 'c2' in messages
 
     # block FF
-    m2 = repo.make_commit('heads/master', 'NO!', None, tree={'m': 'm2'})
+    with repo:
+        m2 = repo.make_commit('heads/master', 'NO!', None, tree={'m': 'm2'})
 
     old_staging = repo.commit('heads/staging.master')
     # confirm staging
-    repo.post_status('heads/staging.master', 'success', 'legal/cla')
-    repo.post_status('heads/staging.master', 'success', 'ci/runbot')
-    run_crons(env)
+    with repo:
+        repo.post_status('heads/staging.master', 'success', 'legal/cla')
+        repo.post_status('heads/staging.master', 'success', 'ci/runbot')
+    env.run_crons()
     new_staging = repo.commit('heads/staging.master')
 
     assert new_staging.id != old_staging.id
 
     # confirm again
-    repo.post_status('heads/staging.master', 'success', 'legal/cla')
-    repo.post_status('heads/staging.master', 'success', 'ci/runbot')
-    run_crons(env)
+    with repo:
+        repo.post_status('heads/staging.master', 'success', 'legal/cla')
+        repo.post_status('heads/staging.master', 'success', 'ci/runbot')
+    env.run_crons()
     messages = {
         c['commit']['message']
         for c in repo.log('heads/master')
@@ -599,30 +637,31 @@ class TestPREdition:
             'project_id': env['runbot_merge.project'].search([]).id,
         })
 
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        repo.make_ref('heads/master', m)
-        repo.make_ref('heads/1.0', m)
-        repo.make_ref('heads/2.0', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            repo.make_ref('heads/master', m)
+            repo.make_ref('heads/1.0', m)
+            repo.make_ref('heads/2.0', m)
 
-        c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
-        c2 = repo.make_commit(c1, 'second', None, tree={'m': 'c2'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c2, user='user')
+            c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
+            c2 = repo.make_commit(c1, 'second', None, tree={'m': 'c2'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=c2)
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
         ])
         assert pr.message == 'title\n\nbody'
-        prx.title = "title 2"
+        with repo: prx.title = "title 2"
         assert pr.message == 'title 2\n\nbody'
-        prx.base = '1.0'
+        with repo: prx.base = '1.0'
         assert pr.target == branch_1
 
-        prx.base = '2.0'
+        with repo: prx.base = '2.0'
         assert not pr.exists()
-        run_crons(env)
+        env.run_crons()
         assert prx.labels == set()
 
-        prx.base = '1.0'
+        with repo: prx.base = '1.0'
         assert env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
@@ -637,26 +676,29 @@ class TestPREdition:
         })
         master = env['runbot_merge.branch'].search([('name', '=', 'master')])
 
-        # master is 1 commit in advance of 1.0
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        m2 = repo.make_commit(m, 'second', None, tree={'m': 'm2'})
-        repo.make_ref('heads/master', m2)
-        repo.make_ref('heads/1.0', m)
+        with repo:
+            # master is 1 commit in advance of 1.0
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            m2 = repo.make_commit(m, 'second', None, tree={'m': 'm2'})
+            repo.make_ref('heads/master', m2)
+            repo.make_ref('heads/1.0', m)
 
-        # the PR builds on master, but is errorneously targeted to 1.0
-        c = repo.make_commit(m2, 'first', None, tree={'m': 'm3'})
-        prx = repo.make_pr('title', 'body', target='1.0', ctid=c, user='user')
+            # the PR builds on master, but is errorneously targeted to 1.0
+            c = repo.make_commit(m2, 'first', None, tree={'m': 'm3'})
+            prx = repo.make_pr(title='title', body='body', target='1.0', head=c)
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
         ])
         assert not pr.squash
 
-        prx.base = 'master'
+        with repo:
+            prx.base = 'master'
         assert pr.target == master
         assert pr.squash
 
-        prx.base = '1.0'
+        with repo:
+            prx.base = '1.0'
         assert pr.target == branch_1
         assert not pr.squash
 
@@ -672,16 +714,17 @@ class TestPREdition:
             'project_id': env['runbot_merge.project'].search([]).id,
         })
 
-        c0 = repo.make_commit(None, '0', None, tree={'a': '0'})
-        repo.make_ref('heads/1.0', c0)
-        c1 = repo.make_commit(c0, '1', None, tree={'a': '1'})
-        repo.make_ref('heads/2.0', c1)
-        c2 = repo.make_commit(c1, '2', None, tree={'a': '2'})
-        repo.make_ref('heads/master', c2)
+        with repo:
+            c0 = repo.make_commit(None, '0', None, tree={'a': '0'})
+            repo.make_ref('heads/1.0', c0)
+            c1 = repo.make_commit(c0, '1', None, tree={'a': '1'})
+            repo.make_ref('heads/2.0', c1)
+            c2 = repo.make_commit(c1, '2', None, tree={'a': '2'})
+            repo.make_ref('heads/master', c2)
 
-        # create PR on 1.0
-        c = repo.make_commit(c0, 'c', None, tree={'a': '0', 'b': '0'})
-        prx = repo.make_pr('t', 'b', target='1.0', ctid=c, user='user')
+            # create PR on 1.0
+            c = repo.make_commit(c0, 'c', None, tree={'a': '0', 'b': '0'})
+            prx = repo.make_pr(title='t', body='b', target='1.0', head=c)
         # there should only be a single PR in the system at this point
         [pr] = env['runbot_merge.pull_requests'].search([])
         assert pr.target == branch_1
@@ -689,9 +732,10 @@ class TestPREdition:
         # branch 1 is EOL, disable it
         branch_1.active = False
 
-        # we forgot we had active PRs for it, and we may want to merge them
-        # still, retarget them!
-        prx.base = '2.0'
+        with repo:
+            # we forgot we had active PRs for it, and we may want to merge them
+            # still, retarget them!
+            prx.base = '2.0'
 
         # check that we still only have one PR in the system
         [pr_] = env['runbot_merge.pull_requests'].search([])
@@ -704,55 +748,60 @@ def test_edit_staged(env, repo):
     """
     What should happen when editing the PR/metadata (not pushing) of a staged PR
     """
-def test_close_staged(env, repo):
+def test_close_staged(env, repo, config):
     """
     When closing a staged PR, cancel the staging
     """
-    m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-    repo.make_ref('heads/master', m)
+    with repo:
+        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+        repo.make_ref('heads/master', m)
 
-    c = repo.make_commit(m, 'fist', None, tree={'m': 'c1'})
-    prx = repo.make_pr('title', 'body', target='master', ctid=c, user='user')
-    repo.post_status(prx.head, 'success', 'legal/cla')
-    repo.post_status(prx.head, 'success', 'ci/runbot')
-    prx.post_comment('hansen r+', user='reviewer')
+        c = repo.make_commit(m, 'fist', None, tree={'m': 'c1'})
+        prx = repo.make_pr(title='title', body='body', target='master', head=c)
+        repo.post_status(prx.head, 'success', 'legal/cla')
+        repo.post_status(prx.head, 'success', 'ci/runbot')
+        prx.post_comment('hansen r+', config['role_reviewer']['token'])
     pr = env['runbot_merge.pull_requests'].search([
         ('repository.name', '=', repo.name),
         ('number', '=', prx.number),
     ])
-    run_crons(env)
+    env.run_crons()
     assert pr.state == 'ready'
     assert pr.staging_id
 
-    prx.close()
-    run_crons(env)
+    with repo:
+        prx.close()
+    env.run_crons()
 
     assert not pr.staging_id
     assert not env['runbot_merge.stagings'].search([])
     assert pr.state == 'closed'
     assert prx.labels == {'seen 🙂', 'closed 💔'}
 
-def test_forward_port(env, repo):
-    m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-    repo.make_ref('heads/master', m)
+def test_forward_port(env, repo, config):
+    with repo:
+        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+        repo.make_ref('heads/master', m)
 
-    head = m
-    for i in range(110):
-        head = repo.make_commit(head, 'c_%03d' % i, None, tree={'m': 'm', 'f': str(i)}, wait=False)
-    # for remote since we're not waiting in commit creation
-    time.sleep(10)
-    pr = repo.make_pr('PR', None, target='master', ctid=head, user='user')
-    repo.post_status(pr.head, 'success', 'legal/cla')
-    repo.post_status(pr.head, 'success', 'ci/runbot')
-    pr.post_comment('hansen r+ merge', "reviewer")
-    run_crons(env)
+        head = m
+        for i in range(110):
+            head = repo.make_commit(head, 'c_%03d' % i, None, tree={'m': 'm', 'f': str(i)})
+    # not sure why we wanted to wait here
+
+    with repo:
+        pr = repo.make_pr(title='PR', body=None, target='master', head=head)
+        repo.post_status(pr.head, 'success', 'legal/cla')
+        repo.post_status(pr.head, 'success', 'ci/runbot')
+        pr.post_comment('hansen r+ merge', config['role_reviewer']['token'])
+    env.run_crons()
 
     st = repo.commit('heads/staging.master')
     assert st.message.startswith('force rebuild')
 
-    repo.post_status(st.id, 'success', 'legal/cla')
-    repo.post_status(st.id, 'success', 'ci/runbot')
-    run_crons(env)
+    with repo:
+        repo.post_status(st.id, 'success', 'legal/cla')
+        repo.post_status(st.id, 'success', 'ci/runbot')
+    env.run_crons()
 
     h = repo.commit('heads/master')
     assert set(st.parents) == {h.id}
@@ -760,7 +809,8 @@ def test_forward_port(env, repo):
     commits = {c['sha'] for c in repo.log('heads/master')}
     assert len(commits) == 112
 
-def test_rebase_failure(env, repo, users, remote_p):
+@pytest.skip("Needs to find a way to make set_ref fail on *second* call.")
+def test_rebase_failure(env, repo, users, config):
     """ It looks like gh.rebase() can fail in the final ref-setting after
     the merging & commits creation has been performed. At this point, the
     staging will fail (yay) but the target branch (tmp) would not get reset,
@@ -781,24 +831,23 @@ def test_rebase_failure(env, repo, users, remote_p):
     but only the first time, and not the set_ref in try_staging itself, and
     that call is performed *in a subprocess* when running <remote> tests.
     """
-    # FIXME: remote mode
-    if remote_p:
-        pytest.skip("Needs to find a way to make set_ref fail on *second* call in remote mode.")
+    with repo:
+        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+        repo.make_ref('heads/master', m)
 
-    m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-    repo.make_ref('heads/master', m)
+        commit_a = repo.make_commit(m, 'A', None, tree={'m': 'm', 'a': 'a'})
+        repo.make_ref('heads/a', commit_a)
+        pr_a = repo.make_pr(title='A', body=None, target='master', head='a')
+        repo.post_status(pr_a.head, 'success', 'ci/runbot')
+        repo.post_status(pr_a.head, 'success', 'legal/cla')
+        pr_a.post_comment('hansen r+', config['role_reviewer']['token'])
 
-    commit_a = repo.make_commit(m, 'A', None, tree={'m': 'm', 'a': 'a'})
-    pr_a = repo.make_pr('A', None, target='master', ctid=commit_a, user='user', label='a')
-    repo.post_status(pr_a.head, 'success', 'ci/runbot')
-    repo.post_status(pr_a.head, 'success', 'legal/cla')
-    pr_a.post_comment('hansen r+', 'reviewer')
-
-    commit_b = repo.make_commit(m, 'B', None, tree={'m': 'm', 'b': 'b'})
-    pr_b = repo.make_pr('B', None, target='master', ctid=commit_b, user='user', label='b')
-    repo.post_status(pr_b.head, 'success', 'ci/runbot')
-    repo.post_status(pr_b.head, 'success', 'legal/cla')
-    pr_b.post_comment('hansen r+', 'reviewer')
+        commit_b = repo.make_commit(m, 'B', None, tree={'m': 'm', 'b': 'b'})
+        repo.make_ref('heads/b', commit_b)
+        pr_b = repo.make_pr(title='B', body=None, target='master', head='b')
+        repo.post_status(pr_b.head, 'success', 'ci/runbot')
+        repo.post_status(pr_b.head, 'success', 'legal/cla')
+        pr_b.post_comment('hansen r+', config['role_reviewer']['token'])
 
     from odoo.addons.runbot_merge.github import GH
     original = GH.set_ref
@@ -825,19 +874,21 @@ def test_rebase_failure(env, repo, users, remote_p):
         'b': 'b',
     }
 
-def test_ci_failure_after_review(env, repo, users):
+def test_ci_failure_after_review(env, repo, users, config):
     """ If a PR is r+'d but the CI ends up failing afterwards, ping the user
     so they're aware. This is useful for the more "fire and forget" approach
     especially small / simple PRs where you assume they're going to pass and
     just r+ immediately.
     """
-    prx = _simple_init(repo)
-    prx.post_comment('hansen r+', "reviewer")
-    run_crons(env)
+    with repo:
+        prx = _simple_init(repo)
+        prx.post_comment('hansen r+', config['role_reviewer']['token'])
+    env.run_crons()
 
-    repo.post_status(prx.head, 'failure', 'ci/runbot')
-    repo.post_status(prx.head, 'success', 'legal/cla')
-    run_crons(env)
+    with repo:
+        repo.post_status(prx.head, 'failure', 'ci/runbot')
+        repo.post_status(prx.head, 'success', 'legal/cla')
+    env.run_crons()
 
     assert prx.comments == [
         (users['reviewer'], 'hansen r+'),
@@ -849,13 +900,14 @@ def test_reopen_state(env, repo):
     already a CI+ stored (as the CI might never trigger unless explicitly
     re-requested)
     """
-    m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-    repo.make_ref('heads/master', m)
+    with repo:
+        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+        repo.make_ref('heads/master', m)
 
-    c = repo.make_commit(m, 'fist', None, tree={'m': 'c1'})
-    repo.post_status(c, 'success', 'legal/cla')
-    repo.post_status(c, 'success', 'ci/runbot')
-    prx = repo.make_pr('title', 'body', target='master', ctid=c, user='user')
+        c = repo.make_commit(m, 'fist', None, tree={'m': 'c1'})
+        repo.post_status(c, 'success', 'legal/cla')
+        repo.post_status(c, 'success', 'ci/runbot')
+        prx = repo.make_pr(title='title', body='body', target='master', head=c)
 
     pr = env['runbot_merge.pull_requests'].search([
         ('repository.name', '=', repo.name),
@@ -864,25 +916,25 @@ def test_reopen_state(env, repo):
     assert pr.state == 'validated', \
         "if a PR is created on a CI'd commit, it should be validated immediately"
 
-    prx.close()
+    with repo: prx.close()
     assert pr.state == 'closed'
 
-    prx.open()
+    with repo: prx.open()
     assert pr.state == 'validated', \
         "if a PR is reopened and had a CI'd head, it should be validated immediately"
 
-def test_no_required_statuses(env, repo):
+def test_no_required_statuses(env, repo, config):
     """ check that mergebot can work on a repo with no CI at all
     """
     env['runbot_merge.project'].search([]).required_statuses = ''
-    m = repo.make_commit(None, 'initial', None, tree={'0': '0'})
-    repo.make_ref('heads/master', m)
+    with repo:
+        m = repo.make_commit(None, 'initial', None, tree={'0': '0'})
+        repo.make_ref('heads/master', m)
 
-    c = repo.make_commit(m, 'first', None, tree={'0': '1'})
-    prx = repo.make_pr('title', 'body', target='master', ctid=c, user='user')
-    prx.post_comment('hansen r+', 'reviewer')
-
-    run_crons(env)
+        c = repo.make_commit(m, 'first', None, tree={'0': '1'})
+        prx = repo.make_pr(title='title', body='body', target='master', head=c)
+        prx.post_comment('hansen r+', config['role_reviewer']['token'])
+    env.run_crons()
 
     assert env['runbot_merge.pull_requests'].search([
         ('repository.name', '=', repo.name),
@@ -891,12 +943,12 @@ def test_no_required_statuses(env, repo):
 
 class TestRetry:
     @pytest.mark.xfail(reason="This may not be a good idea as it could lead to tons of rebuild spam")
-    def test_auto_retry_push(self, env, repo):
+    def test_auto_retry_push(self, env, repo, config):
         prx = _simple_init(repo)
         repo.post_status(prx.head, 'success', 'ci/runbot')
         repo.post_status(prx.head, 'success', 'legal/cla')
-        prx.post_comment('hansen r+', "reviewer")
-        run_crons(env)
+        prx.post_comment('hansen r+', config['role_reviewer']['token'])
+        env.run_crons()
         assert env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
@@ -905,78 +957,84 @@ class TestRetry:
         staging_head = repo.commit('heads/staging.master')
         repo.post_status(staging_head.id, 'success', 'legal/cla')
         repo.post_status(staging_head.id, 'failure', 'ci/runbot')
-        run_crons(env)
+        env.run_crons()
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
         ])
         assert pr.state == 'error'
 
-        prx.push(repo.make_commit(prx.head, 'third', None, tree={'m': 'c3'}))
+        repo.update_ref(prx.ref, repo.make_commit(prx.head, 'third', None, tree={'m': 'c3'}), force=True)
         assert pr.state == 'approved'
         env['runbot_merge.project']._check_progress()
         assert pr.state == 'approved'
         repo.post_status(prx.head, 'success', 'ci/runbot')
         repo.post_status(prx.head, 'success', 'legal/cla')
-        run_crons(env)
+        env.run_crons()
         assert pr.state == 'ready'
 
         staging_head2 = repo.commit('heads/staging.master')
         assert staging_head2 != staging_head
         repo.post_status(staging_head2.id, 'success', 'legal/cla')
         repo.post_status(staging_head2.id, 'success', 'ci/runbot')
-        run_crons(env)
+        env.run_crons()
         assert pr.state == 'merged'
 
     @pytest.mark.parametrize('retrier', ['user', 'other', 'reviewer'])
-    def test_retry_comment(self, env, repo, retrier, users):
+    def test_retry_comment(self, env, repo, retrier, users, config):
         """ An accepted but failed PR should be re-tried when the author or a
         reviewer asks for it
         """
-        prx = _simple_init(repo)
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        prx.post_comment('hansen r+ delegate=%s rebase-merge' % users['other'], "reviewer")
-        run_crons(env)
+        with repo:
+            prx = _simple_init(repo)
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            prx.post_comment('hansen r+ delegate=%s rebase-merge' % users['other'],
+                             config["role_reviewer"]['token'])
+        env.run_crons()
         assert env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
         ]).staging_id
 
         staging_head = repo.commit('heads/staging.master')
-        repo.post_status(staging_head.id, 'success', 'legal/cla')
-        repo.post_status(staging_head.id, 'failure', 'ci/runbot')
-        run_crons(env)
+        with repo:
+            repo.post_status(staging_head.id, 'success', 'legal/cla')
+            repo.post_status(staging_head.id, 'failure', 'ci/runbot')
+        env.run_crons()
         assert env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
         ]).state == 'error'
 
-        prx.post_comment('hansen retry', retrier)
+        with repo:
+            prx.post_comment('hansen retry', config['role_' + retrier]['token'])
         assert env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
         ]).state == 'ready'
-        env['runbot_merge.project']._check_progress()
+        env.run_crons('runbot_merge.merge_cron')
 
         staging_head2 = repo.commit('heads/staging.master')
         assert staging_head2 != staging_head
-        repo.post_status(staging_head2.id, 'success', 'legal/cla')
-        repo.post_status(staging_head2.id, 'success', 'ci/runbot')
-        run_crons(env)
+        with repo:
+            repo.post_status(staging_head2.id, 'success', 'legal/cla')
+            repo.post_status(staging_head2.id, 'success', 'ci/runbot')
+        env.run_crons()
         assert env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
         ]).state == 'merged'
 
-    def test_retry_ignored(self, env, repo, users):
+    def test_retry_ignored(self, env, repo, users, config):
         """ Check feedback in case of ignored retry command on a non-error PR.
         """
-        prx = _simple_init(repo)
-        prx.post_comment('hansen r+', 'reviewer')
-        prx.post_comment('hansen retry', 'reviewer')
+        with repo:
+            prx = _simple_init(repo)
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
+            prx.post_comment('hansen retry', config['role_reviewer']['token'])
+        env.run_crons()
 
-        run_crons(env)
         assert prx.comments == [
             (users['reviewer'], 'hansen r+'),
             (users['reviewer'], 'hansen retry'),
@@ -984,33 +1042,40 @@ class TestRetry:
         ]
 
     @pytest.mark.parametrize('disabler', ['user', 'other', 'reviewer'])
-    def test_retry_disable(self, env, repo, disabler, users):
-        prx = _simple_init(repo)
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        prx.post_comment('hansen r+ delegate=%s rebase-merge' % users['other'], "reviewer")
-        run_crons(env)
+    def test_retry_disable(self, env, repo, disabler, users, config):
+        with repo:
+            prx = _simple_init(repo)
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            prx.post_comment('hansen r+ delegate=%s rebase-merge' % users['other'],
+                             config["role_reviewer"]['token'])
+        env.run_crons()
         assert env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
         ]).staging_id
 
         staging_head = repo.commit('heads/staging.master')
-        repo.post_status(staging_head.id, 'success', 'legal/cla')
-        repo.post_status(staging_head.id, 'failure', 'ci/runbot')
-        run_crons(env)
+        with repo:
+            repo.post_status(staging_head.id, 'success', 'legal/cla')
+            repo.post_status(staging_head.id, 'failure', 'ci/runbot')
+        env.run_crons()
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
         ])
         assert pr.state == 'error'
 
-        prx.post_comment('hansen r-', user=disabler)
+        with repo:
+            prx.post_comment('hansen r-', config['role_' + disabler]['token'])
         assert pr.state == 'validated'
-        prx.push(repo.make_commit(prx.head, 'third', None, tree={'m': 'c3'}))
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        run_crons(env)
+        with repo:
+            repo.make_commit(prx.ref, 'third', None, tree={'m': 'c3'})
+            # just in case, apparently in some case the first post_status uses the old head...
+        with repo:
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            repo.post_status(prx.head, 'success', 'legal/cla')
+        env.run_crons()
         assert pr.state == 'validated'
 
 class TestMergeMethod:
@@ -1018,24 +1083,25 @@ class TestMergeMethod:
     if event['pull_request']['commits'] == 1, "squash" (/rebase); otherwise
     regular merge
     """
-    def test_pr_single_commit(self, repo, env):
+    def test_pr_single_commit(self, repo, env, config):
         """ If single commit, default to rebase & FF
         """
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
-        repo.make_ref('heads/master', m2)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
+            repo.make_ref('heads/master', m2)
 
-        c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c1, user='user')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        prx.post_comment('hansen r+', "reviewer")
+            c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=c1)
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
         assert env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
         ]).squash
 
-        run_crons(env)
+        env.run_crons()
         assert env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
@@ -1053,9 +1119,10 @@ class TestMergeMethod:
         assert actual.parents == [m2],\
             "dummy commit aside, the previous master's tip should be the sole parent of the staging commit"
 
-        repo.post_status(staging.id, 'success', 'legal/cla')
-        repo.post_status(staging.id, 'success', 'ci/runbot')
-        run_crons(env)
+        with repo:
+            repo.post_status(staging.id, 'success', 'legal/cla')
+            repo.post_status(staging.id, 'success', 'ci/runbot')
+        env.run_crons()
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
@@ -1072,19 +1139,21 @@ class TestMergeMethod:
         If a PR starts with 1 commit and a second commit is added, the PR
         should be unflagged as squash
         """
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
-        repo.make_ref('heads/master', m2)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
+            repo.make_ref('heads/master', m2)
 
-        c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c1, user='user')
+            c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=c1)
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number),
         ])
         assert pr.squash, "a PR with a single commit should be squashed"
 
-        prx.push(repo.make_commit(c1, 'second2', None, tree={'m': 'c2'}))
+        with repo:
+            repo.make_commit(prx.ref, 'second2', None, tree={'m': 'c2'})
         assert not pr.squash, "a PR with a single commit should not be squashed"
 
     def test_pr_reset_to_single_commit(self, repo, env):
@@ -1092,39 +1161,46 @@ class TestMergeMethod:
         If a PR starts at >1 commits and is reset back to 1, the PR should be
         re-flagged as squash
         """
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
-        repo.make_ref('heads/master', m2)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
+            repo.make_ref('heads/master', m2)
 
-        c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
-        c2 = repo.make_commit(c1, 'second2', None, tree={'m': 'c2'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c2, user='user')
+            c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
+            c2 = repo.make_commit(c1, 'second2', None, tree={'m': 'c2'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=c2)
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number),
         ])
         assert not pr.squash, "a PR with a single commit should not be squashed"
 
-        prx.push(repo.make_commit(m, 'fixup', None, tree={'m': 'c2'}))
+        with repo:
+            repo.update_ref(
+                prx.ref,
+                repo.make_commit(m, 'fixup', None, tree={'m': 'c2'}),
+                force=True
+            )
         assert pr.squash, "a PR with a single commit should be squashed"
 
-    def test_pr_no_method(self, repo, env, users):
+    def test_pr_no_method(self, repo, env, users, config):
         """ a multi-repo PR should not be staged by default, should also get
         feedback indicating a merge method is necessary
         """
-        m0 = repo.make_commit(None, 'M0', None, tree={'m': '0'})
-        m1 = repo.make_commit(m0, 'M1', None, tree={'m': '1'})
-        m2 = repo.make_commit(m1, 'M2', None, tree={'m': '2'})
-        repo.make_ref('heads/master', m2)
+        with repo:
+            m0 = repo.make_commit(None, 'M0', None, tree={'m': '0'})
+            m1 = repo.make_commit(m0, 'M1', None, tree={'m': '1'})
+            m2 = repo.make_commit(m1, 'M2', None, tree={'m': '2'})
+            repo.make_ref('heads/master', m2)
 
-        b0 = repo.make_commit(m1, 'B0', None, tree={'m': '1', 'b': '0'})
-        b1 = repo.make_commit(b0, 'B1', None, tree={'m': '1', 'b': '1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=b1, user='user')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        prx.post_comment('hansen r+', "reviewer")
+            b0 = repo.make_commit(m1, 'B0', None, tree={'m': '1', 'b': '0'})
+            b1 = repo.make_commit(b0, 'B1', None, tree={'m': '1', 'b': '1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=b1)
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
+        env.run_crons()
 
-        run_crons(env)
         assert not env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number),
@@ -1140,35 +1216,39 @@ class TestMergeMethod:
 """),
         ]
 
-    def test_pr_method_no_review(self, repo, env, users):
+    def test_pr_method_no_review(self, repo, env, users, config):
         """ Configuring the method should be idependent from the review
         """
-        m0 = repo.make_commit(None, 'M0', None, tree={'m': '0'})
-        m1 = repo.make_commit(m0, 'M1', None, tree={'m': '1'})
-        m2 = repo.make_commit(m1, 'M2', None, tree={'m': '2'})
-        repo.make_ref('heads/master', m2)
+        with repo:
+            m0 = repo.make_commit(None, 'M0', None, tree={'m': '0'})
+            m1 = repo.make_commit(m0, 'M1', None, tree={'m': '1'})
+            m2 = repo.make_commit(m1, 'M2', None, tree={'m': '2'})
+            repo.make_ref('heads/master', m2)
 
-        b0 = repo.make_commit(m1, 'B0', None, tree={'m': '1', 'b': '0'})
-        b1 = repo.make_commit(b0, 'B1', None, tree={'m': '1', 'b': '1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=b1, user='user')
+            b0 = repo.make_commit(m1, 'B0', None, tree={'m': '1', 'b': '0'})
+            b1 = repo.make_commit(b0, 'B1', None, tree={'m': '1', 'b': '1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=b1)
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number),
         ])
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
+        with repo:
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            repo.post_status(prx.head, 'success', 'ci/runbot')
 
-        prx.post_comment('hansen rebase-merge', "reviewer")
+            prx.post_comment('hansen rebase-merge', config['role_reviewer']['token'])
         assert pr.merge_method == 'rebase-merge'
-        run_crons(env)
+        env.run_crons()
 
-        prx.post_comment('hansen merge', "reviewer")
+        with repo:
+            prx.post_comment('hansen merge', config['role_reviewer']['token'])
         assert pr.merge_method == 'merge'
-        run_crons(env)
+        env.run_crons()
 
-        prx.post_comment('hansen rebase-ff', "reviewer")
+        with repo:
+            prx.post_comment('hansen rebase-ff', config['role_reviewer']['token'])
         assert pr.merge_method == 'rebase-ff'
-        run_crons(env)
+        env.run_crons()
 
         assert prx.comments == [
             (users['reviewer'], 'hansen rebase-merge'),
@@ -1179,7 +1259,7 @@ class TestMergeMethod:
             (users['user'], "Merge method set to rebase and fast-forward"),
         ]
 
-    def test_pr_rebase_merge(self, repo, env, users):
+    def test_pr_rebase_merge(self, repo, env, users, config):
         """ test result on rebase-merge
 
         left: PR
@@ -1206,25 +1286,25 @@ class TestMergeMethod:
                                            +----------+ merge |
                                                       +-------+
         """
-        m0 = repo.make_commit(None, 'M0', None, tree={'m': '0'})
-        m1 = repo.make_commit(m0, 'M1', None, tree={'m': '1'})
-        m2 = repo.make_commit(m1, 'M2', None, tree={'m': '2'})
-        repo.make_ref('heads/master', m2)
+        with repo:
+            m0 = repo.make_commit(None, 'M0', None, tree={'m': '0'})
+            m1 = repo.make_commit(m0, 'M1', None, tree={'m': '1'})
+            m2 = repo.make_commit(m1, 'M2', None, tree={'m': '2'})
+            repo.make_ref('heads/master', m2)
 
-        # test commit ordering issue while at it: github sorts commits on
-        # author.date instead of doing so topologically which is absolutely
-        # not what we want
-        committer = {'name': 'a', 'email': 'a', 'date': '2018-10-08T11:48:43Z'}
-        author0 = {'name': 'a', 'email': 'a', 'date': '2018-10-01T14:58:38Z'}
-        author1 = {'name': 'a', 'email': 'a', 'date': '2015-10-01T14:58:38Z'}
-        b0 = repo.make_commit(m1, 'B0', author=author0, committer=committer, tree={'m': '1', 'b': '0'})
-        b1 = repo.make_commit(b0, 'B1', author=author1, committer=committer, tree={'m': '1', 'b': '1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=b1, user='user')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        prx.post_comment('hansen r+ rebase-merge', "reviewer")
-
-        run_crons(env)
+            # test commit ordering issue while at it: github sorts commits on
+            # author.date instead of doing so topologically which is absolutely
+            # not what we want
+            committer = {'name': 'a', 'email': 'a', 'date': '2018-10-08T11:48:43Z'}
+            author0 = {'name': 'a', 'email': 'a', 'date': '2018-10-01T14:58:38Z'}
+            author1 = {'name': 'a', 'email': 'a', 'date': '2015-10-01T14:58:38Z'}
+            b0 = repo.make_commit(m1, 'B0', author=author0, committer=committer, tree={'m': '1', 'b': '0'})
+            b1 = repo.make_commit(b0, 'B1', author=author1, committer=committer, tree={'m': '1', 'b': '1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=b1)
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            prx.post_comment('hansen r+ rebase-merge', config['role_reviewer']['token'])
+        env.run_crons()
 
         # create a dag (msg:str, parents:set) from the log
         staging = log_to_node(repo.log('heads/staging.master'))
@@ -1239,9 +1319,10 @@ class TestMergeMethod:
         expected = (re_matches('^force rebuild'), frozenset([merge_head]))
         assert staging == expected
 
-        repo.post_status('heads/staging.master', 'success', 'legal/cla')
-        repo.post_status('heads/staging.master', 'success', 'ci/runbot')
-        run_crons(env)
+        with repo:
+            repo.post_status('heads/staging.master', 'success', 'legal/cla')
+            repo.post_status('heads/staging.master', 'success', 'ci/runbot')
+        env.run_crons()
 
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
@@ -1264,7 +1345,7 @@ class TestMergeMethod:
         }
         assert r0.parents == [m2]
 
-    def test_pr_rebase_ff(self, repo, env, users):
+    def test_pr_rebase_ff(self, repo, env, users, config):
         """ test result on rebase-merge
 
         left: PR
@@ -1288,19 +1369,19 @@ class TestMergeMethod:
           PR |  B1  |                   |  B1  |
              +------+                   +--^---+
         """
-        m0 = repo.make_commit(None, 'M0', None, tree={'m': '0'})
-        m1 = repo.make_commit(m0, 'M1', None, tree={'m': '1'})
-        m2 = repo.make_commit(m1, 'M2', None, tree={'m': '2'})
-        repo.make_ref('heads/master', m2)
+        with repo:
+            m0 = repo.make_commit(None, 'M0', None, tree={'m': '0'})
+            m1 = repo.make_commit(m0, 'M1', None, tree={'m': '1'})
+            m2 = repo.make_commit(m1, 'M2', None, tree={'m': '2'})
+            repo.make_ref('heads/master', m2)
 
-        b0 = repo.make_commit(m1, 'B0', None, tree={'m': '1', 'b': '0'})
-        b1 = repo.make_commit(b0, 'B1', None, tree={'m': '1', 'b': '1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=b1, user='user')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        prx.post_comment('hansen r+ rebase-ff', "reviewer")
-
-        run_crons(env)
+            b0 = repo.make_commit(m1, 'B0', None, tree={'m': '1', 'b': '0'})
+            b1 = repo.make_commit(b0, 'B1', None, tree={'m': '1', 'b': '1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=b1)
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            prx.post_comment('hansen r+ rebase-ff', config['role_reviewer']['token'])
+        env.run_crons()
 
         # create a dag (msg:str, parents:set) from the log
         staging = log_to_node(repo.log('heads/staging.master'))
@@ -1312,9 +1393,10 @@ class TestMergeMethod:
         expected = node(re_matches('^force rebuild'), nb1)
         assert staging == expected
 
-        repo.post_status('heads/staging.master', 'success', 'legal/cla')
-        repo.post_status('heads/staging.master', 'success', 'ci/runbot')
-        run_crons(env)
+        with repo:
+            repo.post_status('heads/staging.master', 'success', 'legal/cla')
+            repo.post_status('heads/staging.master', 'success', 'ci/runbot')
+        env.run_crons()
 
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
@@ -1342,7 +1424,7 @@ class TestMergeMethod:
     def test_pr_contains_merges(self, repo, env):
         pass
 
-    def test_pr_force_merge_single_commit(self, repo, env, users):
+    def test_pr_force_merge_single_commit(self, repo, env, users, config):
         """ should be possible to flag a PR as regular-merged, regardless of
         its commits count
 
@@ -1353,21 +1435,24 @@ class TestMergeMethod:
         C0     +   |
                gib-+
         """
-        m = repo.make_commit(None, "M", None, tree={'a': 'a'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, "M", None, tree={'a': 'a'})
+            repo.make_ref('heads/master', m)
 
-        c0 = repo.make_commit(m, 'C0', None, tree={'a': 'b'})
-        prx = repo.make_pr("gibberish", "blahblah", target='master', ctid=c0, user='user')
-        env['runbot_merge.project']._check_progress()
+            c0 = repo.make_commit(m, 'C0', None, tree={'a': 'b'})
+            prx = repo.make_pr(title="gibberish", body="blahblah", target='master', head=c0)
+        env.run_crons('runbot_merge.merge_cron')
 
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        prx.post_comment('hansen r+ merge', 'reviewer')
-        run_crons(env)
+        with repo:
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            prx.post_comment('hansen r+ merge', config['role_reviewer']['token'])
+        env.run_crons()
 
-        repo.post_status('heads/staging.master', 'success', 'ci/runbot')
-        repo.post_status('heads/staging.master', 'success', 'legal/cla')
-        run_crons(env)
+        with repo:
+            repo.post_status('heads/staging.master', 'success', 'ci/runbot')
+            repo.post_status('heads/staging.master', 'success', 'legal/cla')
+        env.run_crons()
 
         master = repo.commit('heads/master')
         assert master.parents == [m, prx.head], \
@@ -1388,25 +1473,28 @@ class TestMergeMethod:
             '': master.id
         }
 
-    def test_unrebase_emptymessage(self, repo, env, users):
+    def test_unrebase_emptymessage(self, repo, env, users, config):
         """ When merging between master branches (e.g. forward port), the PR
         may have only a title
         """
-        m = repo.make_commit(None, "M", None, tree={'a': 'a'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, "M", None, tree={'a': 'a'})
+            repo.make_ref('heads/master', m)
 
-        c0 = repo.make_commit(m, 'C0', None, tree={'a': 'b'})
-        prx = repo.make_pr("gibberish", None, target='master', ctid=c0, user='user')
-        env['runbot_merge.project']._check_progress()
+            c0 = repo.make_commit(m, 'C0', None, tree={'a': 'b'})
+            prx = repo.make_pr(title="gibberish", body=None, target='master', head=c0)
+        env.run_crons('runbot_merge.merge_cron')
 
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        prx.post_comment('hansen r+ merge', 'reviewer')
-        run_crons(env)
+        with repo:
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            prx.post_comment('hansen r+ merge', config['role_reviewer']['token'])
+        env.run_crons()
 
-        repo.post_status('heads/staging.master', 'success', 'ci/runbot')
-        repo.post_status('heads/staging.master', 'success', 'legal/cla')
-        run_crons(env)
+        with repo:
+            repo.post_status('heads/staging.master', 'success', 'ci/runbot')
+            repo.post_status('heads/staging.master', 'success', 'legal/cla')
+        env.run_crons()
 
         master = repo.commit('heads/master')
         assert master.parents == [m, prx.head], \
@@ -1419,7 +1507,7 @@ class TestMergeMethod:
                         '\n\nSigned-off-by: {}'.format(repo.name, prx.number, reviewer), m, c0)
         assert log_to_node(repo.log('heads/master')), expected
 
-    def test_pr_mergehead(self, repo, env):
+    def test_pr_mergehead(self, repo, env, config):
         """ if the head of the PR is a merge commit and one of the parents is
         in the target, replicate the merge commit instead of merging
 
@@ -1431,23 +1519,26 @@ class TestMergeMethod:
 
         C1 [label = "\\N / MERGE"]
         """
-        m1 = repo.make_commit(None, "M1", None, tree={'a': '0'})
-        m2 = repo.make_commit(m1, "M2", None, tree={'a': '1'})
-        repo.make_ref('heads/master', m2)
+        with repo:
+            m1 = repo.make_commit(None, "M1", None, tree={'a': '0'})
+            m2 = repo.make_commit(m1, "M2", None, tree={'a': '1'})
+            repo.make_ref('heads/master', m2)
 
-        c0 = repo.make_commit(m1, 'C0', None, tree={'a': '0', 'b': '2'})
-        c1 = repo.make_commit([c0, m2], 'C1', None, tree={'a': '1', 'b': '2'})
-        prx = repo.make_pr("T", "TT", target='master', ctid=c1, user='user')
-        run_crons(env)
+            c0 = repo.make_commit(m1, 'C0', None, tree={'a': '0', 'b': '2'})
+            c1 = repo.make_commit([c0, m2], 'C1', None, tree={'a': '1', 'b': '2'})
+            prx = repo.make_pr(title="T", body="TT", target='master', head=c1)
+        env.run_crons()
 
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        prx.post_comment('hansen r+ merge', 'reviewer')
-        run_crons(env)
+        with repo:
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            prx.post_comment('hansen r+ merge', config['role_reviewer']['token'])
+        env.run_crons()
 
-        repo.post_status('heads/staging.master', 'success', 'ci/runbot')
-        repo.post_status('heads/staging.master', 'success', 'legal/cla')
-        run_crons(env)
+        with repo:
+            repo.post_status('heads/staging.master', 'success', 'ci/runbot')
+            repo.post_status('heads/staging.master', 'success', 'legal/cla')
+        env.run_crons()
 
         master = repo.commit('heads/master')
         assert master.parents == [m2, c0]
@@ -1455,7 +1546,7 @@ class TestMergeMethod:
         expected = node('C1', node('C0', m1), node('M2', m1))
         assert log_to_node(repo.log('heads/master')), expected
 
-    def test_pr_mergehead_nonmember(self, repo, env, users):
+    def test_pr_mergehead_nonmember(self, repo, env, users, config):
         """ if the head of the PR is a merge commit but none of the parents is
         in the target, merge normally
 
@@ -1469,25 +1560,28 @@ class TestMergeMethod:
         MERGE -> M2
         MERGE -> C1
         """
-        m1 = repo.make_commit(None, "M1", None, tree={'a': '0'})
-        m2 = repo.make_commit(m1, "M2", None, tree={'a': '1'})
-        repo.make_ref('heads/master', m2)
+        with repo:
+            m1 = repo.make_commit(None, "M1", None, tree={'a': '0'})
+            m2 = repo.make_commit(m1, "M2", None, tree={'a': '1'})
+            repo.make_ref('heads/master', m2)
 
-        b0 = repo.make_commit(m1, 'B0', None, tree={'a': '0', 'bb': 'bb'})
+            b0 = repo.make_commit(m1, 'B0', None, tree={'a': '0', 'bb': 'bb'})
 
-        c0 = repo.make_commit(m1, 'C0', None, tree={'a': '0', 'b': '2'})
-        c1 = repo.make_commit([c0, b0], 'C1', None, tree={'a': '0', 'b': '2', 'bb': 'bb'})
-        prx = repo.make_pr("T", "TT", target='master', ctid=c1, user='user')
-        run_crons(env)
+            c0 = repo.make_commit(m1, 'C0', None, tree={'a': '0', 'b': '2'})
+            c1 = repo.make_commit([c0, b0], 'C1', None, tree={'a': '0', 'b': '2', 'bb': 'bb'})
+            prx = repo.make_pr(title="T", body="TT", target='master', head=c1)
+        env.run_crons()
 
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        prx.post_comment('hansen r+ merge', 'reviewer')
-        run_crons(env)
+        with repo:
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            prx.post_comment('hansen r+ merge', config['role_reviewer']['token'])
+        env.run_crons()
 
-        repo.post_status('heads/staging.master', 'success', 'ci/runbot')
-        repo.post_status('heads/staging.master', 'success', 'legal/cla')
-        run_crons(env)
+        with repo:
+            repo.post_status('heads/staging.master', 'success', 'ci/runbot')
+            repo.post_status('heads/staging.master', 'success', 'legal/cla')
+        env.run_crons()
 
         master = repo.commit('heads/master')
         assert master.parents == [m2, c1]
@@ -1503,23 +1597,23 @@ class TestMergeMethod:
         assert log_to_node(repo.log('heads/master')), expected
 
     @pytest.mark.xfail(reason="removed support for squash+ command")
-    def test_force_squash_merge(self, repo, env):
+    def test_force_squash_merge(self, repo, env, config):
         m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
         m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
         repo.make_ref('heads/master', m2)
 
         c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
         c2 = repo.make_commit(c1, 'second', None, tree={'m': 'c2'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c2, user='user')
+        prx = repo.make_pr(title='title', body='body', target='master', head=c2)
         repo.post_status(prx.head, 'success', 'legal/cla')
         repo.post_status(prx.head, 'success', 'ci/runbot')
-        prx.post_comment('hansen r+ squash+', "reviewer")
+        prx.post_comment('hansen r+ squash+', config['role_reviewer']['token'])
         assert env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
         ]).squash
 
-        run_crons(env)
+        env.run_crons()
         assert env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
@@ -1536,7 +1630,7 @@ class TestMergeMethod:
 
         repo.post_status(staging.id, 'success', 'legal/cla')
         repo.post_status(staging.id, 'success', 'ci/runbot')
-        run_crons(env)
+        env.run_crons()
         assert env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
@@ -1544,22 +1638,22 @@ class TestMergeMethod:
         assert prx.state == 'closed'
 
     @pytest.mark.xfail(reason="removed support for squash- command")
-    def test_disable_squash_merge(self, repo, env):
+    def test_disable_squash_merge(self, repo, env, config):
         m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
         m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
         repo.make_ref('heads/master', m2)
 
         c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c1, user='user')
+        prx = repo.make_pr(title='title', body='body', target='master', head=c1)
         repo.post_status(prx.head, 'success', 'legal/cla')
         repo.post_status(prx.head, 'success', 'ci/runbot')
-        prx.post_comment('hansen r+ squash-', "reviewer")
+        prx.post_comment('hansen r+ squash-', config['role_reviewer']['token'])
         assert not env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
         ]).squash
 
-        run_crons(env)
+        env.run_crons()
         assert env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
@@ -1574,7 +1668,7 @@ class TestMergeMethod:
 
         repo.post_status(staging.id, 'success', 'legal/cla')
         repo.post_status(staging.id, 'success', 'ci/runbot')
-        run_crons(env)
+        env.run_crons()
         assert env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
@@ -1586,53 +1680,60 @@ class TestPRUpdate(object):
     can have additional effect (see individual tests)
     """
     def test_update_opened(self, env, repo):
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            repo.make_ref('heads/master', m)
 
-        c = repo.make_commit(m, 'fist', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c, user='user')
+            c = repo.make_commit(m, 'fist', None, tree={'m': 'c1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=c)
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number),
         ])
         assert pr.head == c
         # alter & push force PR entirely
-        c2 = repo.make_commit(m, 'first', None, tree={'m': 'cc'})
-        prx.push(c2)
+        with repo:
+            c2 = repo.make_commit(m, 'first', None, tree={'m': 'cc'})
+            repo.update_ref(prx.ref, c2, force=True)
         assert pr.head == c2
 
     def test_reopen_update(self, env, repo):
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            repo.make_ref('heads/master', m)
 
-        c = repo.make_commit(m, 'fist', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c, user='user')
+            c = repo.make_commit(m, 'fist', None, tree={'m': 'c1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=c)
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number),
         ])
-        prx.close()
+        with repo:
+            prx.close()
         assert pr.state == 'closed'
         assert pr.head == c
 
-        prx.open()
+        with repo:
+            prx.open()
         assert pr.state == 'opened'
 
-        c2 = repo.make_commit(c, 'first', None, tree={'m': 'cc'})
-        prx.push(c2)
+        with repo:
+            c2 = repo.make_commit(c, 'first', None, tree={'m': 'cc'})
+            repo.update_ref(prx.ref, c2, force=True)
         assert pr.head == c2
 
     def test_update_validated(self, env, repo):
         """ Should reset to opened
         """
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            repo.make_ref('heads/master', m)
 
-        c = repo.make_commit(m, 'fist', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c, user='user')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        run_crons(env)
+            c = repo.make_commit(m, 'fist', None, tree={'m': 'c1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=c)
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+        env.run_crons()
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number),
@@ -1640,18 +1741,20 @@ class TestPRUpdate(object):
         assert pr.head == c
         assert pr.state == 'validated'
 
-        c2 = repo.make_commit(m, 'first', None, tree={'m': 'cc'})
-        prx.push(c2)
+        with repo:
+            c2 = repo.make_commit(m, 'first', None, tree={'m': 'cc'})
+            repo.update_ref(prx.ref, c2, force=True)
         assert pr.head == c2
         assert pr.state == 'opened'
 
-    def test_update_approved(self, env, repo):
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        repo.make_ref('heads/master', m)
+    def test_update_approved(self, env, repo, config):
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            repo.make_ref('heads/master', m)
 
-        c = repo.make_commit(m, 'fist', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c, user='user')
-        prx.post_comment('hansen r+', user='reviewer')
+            c = repo.make_commit(m, 'fist', None, tree={'m': 'c1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=c)
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number),
@@ -1659,23 +1762,25 @@ class TestPRUpdate(object):
         assert pr.head == c
         assert pr.state == 'approved'
 
-        c2 = repo.make_commit(c, 'first', None, tree={'m': 'cc'})
-        prx.push(c2)
+        with repo:
+            c2 = repo.make_commit(c, 'first', None, tree={'m': 'cc'})
+            repo.update_ref(prx.ref, c2, force=True)
         assert pr.head == c2
         assert pr.state == 'opened'
 
-    def test_update_ready(self, env, repo):
+    def test_update_ready(self, env, repo, config):
         """ Should reset to opened
         """
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            repo.make_ref('heads/master', m)
 
-        c = repo.make_commit(m, 'fist', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c, user='user')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        prx.post_comment('hansen r+', user='reviewer')
-        run_crons(env)
+            c = repo.make_commit(m, 'fist', None, tree={'m': 'c1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=c)
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
+        env.run_crons()
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number),
@@ -1683,57 +1788,62 @@ class TestPRUpdate(object):
         assert pr.head == c
         assert pr.state == 'ready'
 
-        c2 = repo.make_commit(c, 'first', None, tree={'m': 'cc'})
-        prx.push(c2)
+        with repo:
+            c2 = repo.make_commit(c, 'first', None, tree={'m': 'cc'})
+            repo.update_ref(prx.ref, c2, force=True)
         assert pr.head == c2
         assert pr.state == 'opened'
 
-    def test_update_staged(self, env, repo):
+    def test_update_staged(self, env, repo, config):
         """ Should cancel the staging & reset PR to opened
         """
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            repo.make_ref('heads/master', m)
 
-        c = repo.make_commit(m, 'fist', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c, user='user')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        prx.post_comment('hansen r+', user='reviewer')
+            c = repo.make_commit(m, 'fist', None, tree={'m': 'c1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=c)
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number),
         ])
-        run_crons(env)
+        env.run_crons()
         assert pr.state == 'ready'
         assert pr.staging_id
 
-        c2 = repo.make_commit(c, 'first', None, tree={'m': 'cc'})
-        prx.push(c2)
+        with repo:
+            c2 = repo.make_commit(c, 'first', None, tree={'m': 'cc'})
+            repo.update_ref(prx.ref, c2, force=True)
         assert pr.head == c2
         assert pr.state == 'opened'
         assert not pr.staging_id
         assert not env['runbot_merge.stagings'].search([])
 
-    def test_split(self, env, repo):
+    def test_split(self, env, repo, config):
         """ Should remove the PR from its split, and possibly delete the split
         entirely.
         """
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            repo.make_ref('heads/master', m)
 
-        c = repo.make_commit(m, 'first', None, tree={'m': 'm', '1': '1'})
-        prx1 = repo.make_pr('t1', 'b1', target='master', ctid=c, user='user', label='p1')
-        repo.post_status(prx1.head, 'success', 'legal/cla')
-        repo.post_status(prx1.head, 'success', 'ci/runbot')
-        prx1.post_comment('hansen r+', user='reviewer')
+            c = repo.make_commit(m, 'first', None, tree={'m': 'm', '1': '1'})
+            repo.make_ref('heads/p1', c)
+            prx1 = repo.make_pr(title='t1', body='b1', target='master', head='p1')
+            repo.post_status(prx1.head, 'success', 'legal/cla')
+            repo.post_status(prx1.head, 'success', 'ci/runbot')
+            prx1.post_comment('hansen r+', config['role_reviewer']['token'])
 
-        c = repo.make_commit(m, 'first', None, tree={'m': 'm', '2': '2'})
-        prx2 = repo.make_pr('t2', 'b2', target='master', ctid=c, user='user', label='p2')
-        repo.post_status(prx2.head, 'success', 'legal/cla')
-        repo.post_status(prx2.head, 'success', 'ci/runbot')
-        prx2.post_comment('hansen r+', user='reviewer')
-
-        run_crons(env)
+            c = repo.make_commit(m, 'first', None, tree={'m': 'm', '2': '2'})
+            repo.make_ref('heads/p2', c)
+            prx2 = repo.make_pr(title='t2', body='b2', target='master', head='p2')
+            repo.post_status(prx2.head, 'success', 'legal/cla')
+            repo.post_status(prx2.head, 'success', 'ci/runbot')
+            prx2.post_comment('hansen r+', config['role_reviewer']['token'])
+        env.run_crons()
 
         pr1, pr2 = env['runbot_merge.pull_requests'].search([], order='number')
         assert pr1.number == prx1.number
@@ -1741,8 +1851,9 @@ class TestPRUpdate(object):
         assert pr1.staging_id == pr2.staging_id
         s0 = pr1.staging_id
 
-        repo.post_status('heads/staging.master', 'failure', 'ci/runbot')
-        run_crons(env)
+        with repo:
+            repo.post_status('heads/staging.master', 'failure', 'ci/runbot')
+        env.run_crons()
 
         assert pr1.staging_id and pr1.staging_id != s0, "pr1 should have been re-staged"
         assert not pr2.staging_id, "pr2 should not"
@@ -1750,56 +1861,62 @@ class TestPRUpdate(object):
         #       around using active_test=False
         assert env['runbot_merge.split'].search([])
 
-        prx2.push(repo.make_commit(c, 'second', None, tree={'m': 'm', '2': '22'}))
+        with repo:
+            repo.update_ref(prx2.ref, repo.make_commit(c, 'second', None, tree={'m': 'm', '2': '22'}), force=True)
         # probably not necessary ATM but...
-        run_crons(env)
+        env.run_crons()
 
         assert pr2.state == 'opened', "state should have been reset"
         assert not env['runbot_merge.split'].search([]), "there should be no split left"
 
-    def test_update_error(self, env, repo):
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        repo.make_ref('heads/master', m)
+    def test_update_error(self, env, repo, config):
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            repo.make_ref('heads/master', m)
 
-        c = repo.make_commit(m, 'fist', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c, user='user')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        prx.post_comment('hansen r+', user='reviewer')
+            c = repo.make_commit(m, 'fist', None, tree={'m': 'c1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=c)
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number),
         ])
-        run_crons(env)
+        env.run_crons()
         assert pr.state == 'ready'
         assert pr.staging_id
 
         h = repo.commit('heads/staging.master').id
-        repo.post_status(h, 'success', 'legal/cla')
-        repo.post_status(h, 'failure', 'ci/runbot')
-        run_crons(env)
+        with repo:
+            repo.post_status(h, 'success', 'legal/cla')
+            repo.post_status(h, 'failure', 'ci/runbot')
+        env.run_crons()
         assert not pr.staging_id
         assert pr.state == 'error'
 
-        c2 = repo.make_commit(c, 'first', None, tree={'m': 'cc'})
-        prx.push(c2)
+        with repo:
+            c2 = repo.make_commit(c, 'first', None, tree={'m': 'cc'})
+            repo.update_ref(prx.ref, c2, force=True)
         assert pr.head == c2
         assert pr.state == 'opened'
 
     def test_unknown_pr(self, env, repo):
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        repo.make_ref('heads/1.0', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            repo.make_ref('heads/1.0', m)
 
-        c = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='1.0', ctid=c, user='user')
+            c = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
+            prx = repo.make_pr(title='title', body='body', target='1.0', head=c)
         assert not env['runbot_merge.pull_requests'].search([('number', '=', prx.number)])
 
         env['runbot_merge.project'].search([]).write({
             'branch_ids': [(0, 0, {'name': '1.0'})]
         })
 
-        c2 = repo.make_commit(c, 'second', None, tree={'m': 'c2'})
-        prx.push(c2)
+        with repo:
+            c2 = repo.make_commit(c, 'second', None, tree={'m': 'c2'})
+            repo.update_ref(prx.ref, c2, force=True)
 
         assert not env['runbot_merge.pull_requests'].search([('number', '=', prx.number)])
 
@@ -1807,16 +1924,18 @@ class TestPRUpdate(object):
         """ If a PR is updated to a known-valid commit, it should be
         validated
         """
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            repo.make_ref('heads/master', m)
 
-        c = repo.make_commit(m, 'fist', None, tree={'m': 'c1'})
-        c2 = repo.make_commit(m, 'first', None, tree={'m': 'cc'})
-        repo.post_status(c2, 'success', 'legal/cla')
-        repo.post_status(c2, 'success', 'ci/runbot')
-        run_crons(env)
+            c = repo.make_commit(m, 'fist', None, tree={'m': 'c1'})
+            c2 = repo.make_commit(m, 'first', None, tree={'m': 'cc'})
+            repo.post_status(c2, 'success', 'legal/cla')
+            repo.post_status(c2, 'success', 'ci/runbot')
+        env.run_crons()
 
-        prx = repo.make_pr('title', 'body', target='master', ctid=c, user='user')
+        with repo:
+            prx = repo.make_pr(title='title', body='body', target='master', head=c)
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number),
@@ -1824,34 +1943,27 @@ class TestPRUpdate(object):
         assert pr.head == c
         assert pr.state == 'opened'
 
-        prx.push(c2)
+        with repo:
+            repo.update_ref(prx.ref, c2, force=True)
         assert pr.head == c2
         assert pr.state == 'validated'
 
 class TestBatching(object):
-    def _pr(self, repo, prefix, trees, *,
-            target='master', user='user', reviewer='reviewer',
+    def _pr(self, repo, prefix, trees, *, target='master', user, reviewer,
             statuses=(('ci/runbot', 'success'), ('legal/cla', 'success'))
         ):
         """ Helper creating a PR from a series of commits on a base
-
-        :type repo: fake_github.Repo
-        :param prefix: a prefix used for commit messages, PR title & PR body
-        :param trees: a list of dicts symbolising the tree for the corresponding commit.
-                      each tree is an update on the "current state" of the tree
-        :param target: branch, both the base commit and the PR target
-        :type target: str
-        :type user: str
-        :type reviewer: str | None
-        :type statuses: List[(str, str)]
         """
-        base = repo.commit('heads/{}'.format(target))
-        tree = repo.read_tree(base)
-        c = base.id
-        for i, t in enumerate(trees):
-            tree.update(t)
-            c = repo.make_commit(c, 'commit_{}_{:02}'.format(prefix, i), None, tree=dict(tree))
-        pr = repo.make_pr('title {}'.format(prefix), 'body {}'.format(prefix), target=target, ctid=c, user=user, label=prefix)
+        *_, c = repo.make_commits(
+            'heads/{}'.format(target),
+            *(
+                repo.Commit('commit_{}_{:02}'.format(prefix, i), tree=t)
+                for i, t in enumerate(trees)
+            ),
+            ref='heads/{}'.format(prefix)
+        )
+        pr = repo.make_pr(title='title {}'.format(prefix), body='body {}'.format(prefix),
+                          target=target, head=prefix, token=user)
 
         for context, result in statuses:
             repo.post_status(c, result, context)
@@ -1868,17 +1980,18 @@ class TestBatching(object):
             ('number', '=', number),
         ])
 
-    def test_staging_batch(self, env, repo, users):
+    def test_staging_batch(self, env, repo, users, config):
         """ If multiple PRs are ready for the same target at the same point,
         they should be staged together
         """
-        m = repo.make_commit(None, 'initial', None, tree={'a': 'some content'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'a': 'some content'})
+            repo.make_ref('heads/master', m)
 
-        pr1 = self._pr(repo, 'PR1', [{'a': 'AAA'}, {'b': 'BBB'}])
-        pr2 = self._pr(repo, 'PR2', [{'c': 'CCC'}, {'d': 'DDD'}])
+            pr1 = self._pr(repo, 'PR1', [{'a': 'AAA'}, {'b': 'BBB'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'])
+            pr2 = self._pr(repo, 'PR2', [{'c': 'CCC'}, {'d': 'DDD'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'])
+        env.run_crons()
 
-        run_crons(env)
         pr1 = self._get(env, repo, pr1.number)
         assert pr1.staging_id
         pr2 = self._get(env, repo, pr2.number)
@@ -1902,19 +2015,20 @@ class TestBatching(object):
         expected = (re_matches('^force rebuild'), frozenset([p2]))
         assert staging == expected
 
-    def test_staging_batch_norebase(self, env, repo, users):
+    def test_staging_batch_norebase(self, env, repo, users, config):
         """ If multiple PRs are ready for the same target at the same point,
         they should be staged together
         """
-        m = repo.make_commit(None, 'initial', None, tree={'a': 'some content'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'a': 'some content'})
+            repo.make_ref('heads/master', m)
 
-        pr1 = self._pr(repo, 'PR1', [{'a': 'AAA'}, {'b': 'BBB'}])
-        pr1.post_comment('hansen merge', 'reviewer')
-        pr2 = self._pr(repo, 'PR2', [{'c': 'CCC'}, {'d': 'DDD'}])
-        pr2.post_comment('hansen merge', 'reviewer')
+            pr1 = self._pr(repo, 'PR1', [{'a': 'AAA'}, {'b': 'BBB'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'])
+            pr1.post_comment('hansen merge', config['role_reviewer']['token'])
+            pr2 = self._pr(repo, 'PR2', [{'c': 'CCC'}, {'d': 'DDD'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'])
+            pr2.post_comment('hansen merge', config['role_reviewer']['token'])
+        env.run_crons()
 
-        run_crons(env)
         pr1 = self._get(env, repo, pr1.number)
         assert pr1.staging_id
         assert pr1.merge_method == 'merge'
@@ -1942,17 +2056,18 @@ class TestBatching(object):
         expected = (re_matches('^force rebuild'), frozenset([p2]))
         assert staging == expected
 
-    def test_staging_batch_squash(self, env, repo, users):
+    def test_staging_batch_squash(self, env, repo, users, config):
         """ If multiple PRs are ready for the same target at the same point,
         they should be staged together
         """
-        m = repo.make_commit(None, 'initial', None, tree={'a': 'some content'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'a': 'some content'})
+            repo.make_ref('heads/master', m)
 
-        pr1 = self._pr(repo, 'PR1', [{'a': 'AAA'}])
-        pr2 = self._pr(repo, 'PR2', [{'c': 'CCC'}])
+            pr1 = self._pr(repo, 'PR1', [{'a': 'AAA'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'])
+            pr2 = self._pr(repo, 'PR2', [{'c': 'CCC'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'])
+        env.run_crons()
 
-        run_crons(env)
         pr1 = self._get(env, repo, pr1.number)
         assert pr1.staging_id
         pr2 = self._get(env, repo, pr2.number)
@@ -1971,25 +2086,26 @@ class TestBatching(object):
                       node('initial'))))
         assert staging == expected
 
-    def test_batching_pressing(self, env, repo):
+    def test_batching_pressing(self, env, repo, config):
         """ "Pressing" PRs should be selected before normal & batched together
         """
-        m = repo.make_commit(None, 'initial', None, tree={'a': 'some content'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'a': 'some content'})
+            repo.make_ref('heads/master', m)
 
-        pr21 = self._pr(repo, 'PR1', [{'a': 'AAA'}, {'b': 'BBB'}])
-        pr22 = self._pr(repo, 'PR2', [{'c': 'CCC'}, {'d': 'DDD'}])
+            pr21 = self._pr(repo, 'PR1', [{'a': 'AAA'}, {'b': 'BBB'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'])
+            pr22 = self._pr(repo, 'PR2', [{'c': 'CCC'}, {'d': 'DDD'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'])
 
-        pr11 = self._pr(repo, 'Pressing1', [{'x': 'x'}, {'y': 'y'}])
-        pr12 = self._pr(repo, 'Pressing2', [{'z': 'z'}, {'zz': 'zz'}])
-        pr11.post_comment('hansen priority=1', 'reviewer')
-        pr12.post_comment('hansen priority=1', 'reviewer')
+            pr11 = self._pr(repo, 'Pressing1', [{'x': 'x'}, {'y': 'y'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'])
+            pr12 = self._pr(repo, 'Pressing2', [{'z': 'z'}, {'zz': 'zz'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'])
+            pr11.post_comment('hansen priority=1', config['role_reviewer']['token'])
+            pr12.post_comment('hansen priority=1', config['role_reviewer']['token'])
 
         pr21, pr22, pr11, pr12 = prs = [self._get(env, repo, pr.number) for pr in [pr21, pr22, pr11, pr12]]
         assert pr21.priority == pr22.priority == 2
         assert pr11.priority == pr12.priority == 1
 
-        run_crons(env)
+        env.run_crons()
 
         assert all(pr.state == 'ready' for pr in prs)
         assert not pr21.staging_id
@@ -1998,20 +2114,21 @@ class TestBatching(object):
         assert pr12.staging_id
         assert pr11.staging_id == pr12.staging_id
 
-    def test_batching_urgent(self, env, repo):
-        m = repo.make_commit(None, 'initial', None, tree={'a': 'some content'})
-        repo.make_ref('heads/master', m)
+    def test_batching_urgent(self, env, repo, config):
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'a': 'some content'})
+            repo.make_ref('heads/master', m)
 
-        pr21 = self._pr(repo, 'PR1', [{'a': 'AAA'}, {'b': 'BBB'}])
-        pr22 = self._pr(repo, 'PR2', [{'c': 'CCC'}, {'d': 'DDD'}])
+            pr21 = self._pr(repo, 'PR1', [{'a': 'AAA'}, {'b': 'BBB'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'])
+            pr22 = self._pr(repo, 'PR2', [{'c': 'CCC'}, {'d': 'DDD'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'])
 
-        pr11 = self._pr(repo, 'Pressing1', [{'x': 'x'}, {'y': 'y'}])
-        pr12 = self._pr(repo, 'Pressing2', [{'z': 'z'}, {'zz': 'zz'}])
-        pr11.post_comment('hansen priority=1', 'reviewer')
-        pr12.post_comment('hansen priority=1', 'reviewer')
+            pr11 = self._pr(repo, 'Pressing1', [{'x': 'x'}, {'y': 'y'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'])
+            pr12 = self._pr(repo, 'Pressing2', [{'z': 'z'}, {'zz': 'zz'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'])
+            pr11.post_comment('hansen priority=1', config['role_reviewer']['token'])
+            pr12.post_comment('hansen priority=1', config['role_reviewer']['token'])
 
         # stage PR1
-        run_crons(env)
+        env.run_crons()
         p_11, p_12, p_21, p_22 = \
             [self._get(env, repo, pr.number) for pr in [pr11, pr12, pr21, pr22]]
         assert not p_21.staging_id or p_22.staging_id
@@ -2020,70 +2137,76 @@ class TestBatching(object):
         staging_1 = p_11.staging_id
 
         # no statuses run on PR0s
-        pr01 = self._pr(repo, 'Urgent1', [{'n': 'n'}, {'o': 'o'}], reviewer=None, statuses=[])
-        pr01.post_comment('hansen priority=0 rebase-merge', 'reviewer')
+        with repo:
+            pr01 = self._pr(repo, 'Urgent1', [{'n': 'n'}, {'o': 'o'}], user=config['role_user']['token'], reviewer=None, statuses=[])
+            pr01.post_comment('hansen priority=0 rebase-merge', config['role_reviewer']['token'])
         p_01 = self._get(env, repo, pr01.number)
         assert p_01.state == 'opened'
         assert p_01.priority == 0
 
-        run_crons(env)
+        env.run_crons()
         # first staging should be cancelled and PR0 should be staged
         # regardless of CI (or lack thereof)
         assert not staging_1.active
         assert not p_11.staging_id and not p_12.staging_id
         assert p_01.staging_id
 
-    def test_batching_urgenter_than_split(self, env, repo):
+    def test_batching_urgenter_than_split(self, env, repo, config):
         """ p=0 PRs should take priority over split stagings (processing
         of a staging having CI-failed and being split into sub-stagings)
         """
-        m = repo.make_commit(None, 'initial', None, tree={'a': 'some content'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'a': 'some content'})
+            repo.make_ref('heads/master', m)
 
-        pr1 = self._pr(repo, 'PR1', [{'a': 'AAA'}, {'b': 'BBB'}])
-        p_1 = self._get(env, repo, pr1.number)
-        pr2 = self._pr(repo, 'PR2', [{'a': 'some content', 'c': 'CCC'}, {'d': 'DDD'}])
-        p_2 = self._get(env, repo, pr2.number)
+            pr1 = self._pr(repo, 'PR1', [{'a': 'AAA'}, {'b': 'BBB'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'])
+            p_1 = self._get(env, repo, pr1.number)
+            pr2 = self._pr(repo, 'PR2', [{'a': 'some content', 'c': 'CCC'}, {'d': 'DDD'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'])
+            p_2 = self._get(env, repo, pr2.number)
+        env.run_crons()
 
-        run_crons(env)
         st = env['runbot_merge.stagings'].search([])
         # both prs should be part of the staging
         assert st.mapped('batch_ids.prs') == p_1 | p_2
         # add CI failure
-        repo.post_status('heads/staging.master', 'failure', 'ci/runbot')
-        repo.post_status('heads/staging.master', 'success', 'legal/cla')
+        with repo:
+            repo.post_status('heads/staging.master', 'failure', 'ci/runbot')
+            repo.post_status('heads/staging.master', 'success', 'legal/cla')
+        env.run_crons()
 
-        run_crons(env)
         # should have staged the first half
         assert p_1.staging_id.heads
         assert not p_2.staging_id.heads
 
         # during restaging of pr1, create urgent PR
-        pr0 = self._pr(repo, 'urgent', [{'a': 'a', 'b': 'b'}], reviewer=None, statuses=[])
-        pr0.post_comment('hansen priority=0', 'reviewer')
+        with repo:
+            pr0 = self._pr(repo, 'urgent', [{'a': 'a', 'b': 'b'}], user=config['role_user']['token'], reviewer=None, statuses=[])
+            pr0.post_comment('hansen priority=0', config['role_reviewer']['token'])
+        env.run_crons()
 
-        run_crons(env)
         # TODO: maybe just deactivate stagings instead of deleting them when canceling?
         assert not p_1.staging_id
         assert self._get(env, repo, pr0.number).staging_id
 
-    def test_urgent_failed(self, env, repo):
+    def test_urgent_failed(self, env, repo, config):
         """ Ensure pr[p=0,state=failed] don't get picked up
         """
-        m = repo.make_commit(None, 'initial', None, tree={'a': 'some content'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'a': 'some content'})
+            repo.make_ref('heads/master', m)
 
-        pr21 = self._pr(repo, 'PR1', [{'a': 'AAA'}, {'b': 'BBB'}])
+            pr21 = self._pr(repo, 'PR1', [{'a': 'AAA'}, {'b': 'BBB'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'])
 
         p_21 = self._get(env, repo, pr21.number)
 
         # no statuses run on PR0s
-        pr01 = self._pr(repo, 'Urgent1', [{'n': 'n'}, {'o': 'o'}], reviewer=None, statuses=[])
-        pr01.post_comment('hansen priority=0', 'reviewer')
+        with repo:
+            pr01 = self._pr(repo, 'Urgent1', [{'n': 'n'}, {'o': 'o'}], user=config['role_user']['token'], reviewer=None, statuses=[])
+            pr01.post_comment('hansen priority=0', config['role_reviewer']['token'])
         p_01 = self._get(env, repo, pr01.number)
         p_01.state = 'error'
 
-        run_crons(env)
+        env.run_crons()
         assert not p_01.staging_id, "p_01 should not be picked up as it's failed"
         assert p_21.staging_id, "p_21 should have been staged"
 
@@ -2091,27 +2214,29 @@ class TestBatching(object):
     def test_batching_merge_failure(self):
         pass
 
-    def test_staging_ci_failure_batch(self, env, repo):
+    def test_staging_ci_failure_batch(self, env, repo, config):
         """ on failure split batch & requeue
         """
-        m = repo.make_commit(None, 'initial', None, tree={'a': 'some content'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'a': 'some content'})
+            repo.make_ref('heads/master', m)
 
-        pr1 = self._pr(repo, 'PR1', [{'a': 'AAA'}, {'b': 'BBB'}])
-        pr2 = self._pr(repo, 'PR2', [{'a': 'some content', 'c': 'CCC'}, {'d': 'DDD'}])
+            pr1 = self._pr(repo, 'PR1', [{'a': 'AAA'}, {'b': 'BBB'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'])
+            pr2 = self._pr(repo, 'PR2', [{'a': 'some content', 'c': 'CCC'}, {'d': 'DDD'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'])
+        env.run_crons()
 
-        run_crons(env)
         st = env['runbot_merge.stagings'].search([])
         # both prs should be part of the staging
         assert len(st.mapped('batch_ids.prs')) == 2
         # add CI failure
-        repo.post_status('heads/staging.master', 'failure', 'ci/runbot')
-        repo.post_status('heads/staging.master', 'success', 'legal/cla')
+        with repo:
+            repo.post_status('heads/staging.master', 'failure', 'ci/runbot')
+            repo.post_status('heads/staging.master', 'success', 'legal/cla')
 
         pr1 = env['runbot_merge.pull_requests'].search([('number', '=', pr1.number)])
         pr2 = env['runbot_merge.pull_requests'].search([('number', '=', pr2.number)])
 
-        run_crons(env)
+        env.run_crons()
         # should have split the existing batch into two, with one of the
         # splits having been immediately restaged
         st = env['runbot_merge.stagings'].search([])
@@ -2123,49 +2248,53 @@ class TestBatching(object):
 
         # This is the failing PR!
         h = repo.commit('heads/staging.master').id
-        repo.post_status(h, 'failure', 'ci/runbot')
-        repo.post_status(h, 'success', 'legal/cla')
-        run_crons(env)
+        with repo:
+            repo.post_status(h, 'failure', 'ci/runbot')
+            repo.post_status(h, 'success', 'legal/cla')
+        env.run_crons()
         assert pr1.state == 'error'
 
         assert pr2.staging_id
 
         h = repo.commit('heads/staging.master').id
-        repo.post_status(h, 'success', 'ci/runbot')
-        repo.post_status(h, 'success', 'legal/cla')
-        env['runbot_merge.commit']._notify()
-        env['runbot_merge.project']._check_progress()
+        with repo:
+            repo.post_status(h, 'success', 'ci/runbot')
+            repo.post_status(h, 'success', 'legal/cla')
+        env.run_crons('runbot_merge.process_updated_commits', 'runbot_merge.merge_cron')
         assert pr2.state == 'merged'
 
 class TestReviewing(object):
-    def test_reviewer_rights(self, env, repo, users):
+    def test_reviewer_rights(self, env, repo, users, config):
         """Only users with review rights will have their r+ (and other
         attributes) taken in account
         """
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
-        repo.make_ref('heads/master', m2)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
+            repo.make_ref('heads/master', m2)
 
-        c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c1, user='user')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        prx.post_comment('hansen r+', user='other')
-        run_crons(env)
+            c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=c1)
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            prx.post_comment('hansen r+', config['role_other']['token'])
+        env.run_crons()
 
         assert env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
         ]).state == 'validated'
-        prx.post_comment('hansen r+', user='reviewer')
+        with repo:
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
         assert env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
         ]).state == 'ready'
         # second r+ to check warning
-        prx.post_comment('hansen r+', user='reviewer')
+        with repo:
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
 
-        run_crons(env)
+        env.run_crons()
         assert prx.comments == [
             (users['other'], 'hansen r+'),
             (users['user'], "I'm sorry, @{}. I'm afraid I can't do that.".format(users['other'])),
@@ -2175,19 +2304,20 @@ class TestReviewing(object):
                  users['reviewer'])),
         ]
 
-    def test_self_review_fail(self, env, repo, users):
+    def test_self_review_fail(self, env, repo, users, config):
         """ Normal reviewers can't self-review
         """
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
-        repo.make_ref('heads/master', m2)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
+            repo.make_ref('heads/master', m2)
 
-        c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c1, user='reviewer')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        prx.post_comment('hansen r+', user='reviewer')
-        run_crons(env)
+            c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=c1, token=config['role_reviewer']['token'])
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
+        env.run_crons()
 
         assert prx.user == users['reviewer']
         assert env['runbot_merge.pull_requests'].search([
@@ -2195,25 +2325,26 @@ class TestReviewing(object):
             ('number', '=', prx.number)
         ]).state == 'validated'
 
-        run_crons(env)
+        env.run_crons()
         assert prx.comments == [
             (users['reviewer'], 'hansen r+'),
             (users['user'], "I'm sorry, @{}. You can't review+.".format(users['reviewer'])),
         ]
 
-    def test_self_review_success(self, env, repo, users):
+    def test_self_review_success(self, env, repo, users, config):
         """ Some users are allowed to self-review
         """
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
-        repo.make_ref('heads/master', m2)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
+            repo.make_ref('heads/master', m2)
 
-        c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c1, user='self_reviewer')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        prx.post_comment('hansen r+', user='self_reviewer')
-        run_crons(env)
+            c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=c1, token=config['role_self_reviewer']['token'])
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            prx.post_comment('hansen r+', config['role_self_reviewer']['token'])
+        env.run_crons()
 
         assert prx.user == users['self_reviewer']
         assert env['runbot_merge.pull_requests'].search([
@@ -2221,21 +2352,22 @@ class TestReviewing(object):
             ('number', '=', prx.number)
         ]).state == 'ready'
 
-    def test_delegate_review(self, env, repo, users):
+    def test_delegate_review(self, env, repo, users, config):
         """Users should be able to delegate review to either the creator of
         the PR or an other user without review rights
         """
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
-        repo.make_ref('heads/master', m2)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
+            repo.make_ref('heads/master', m2)
 
-        c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c1, user='user')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        prx.post_comment('hansen delegate+', user='reviewer')
-        prx.post_comment('hansen r+', user='user')
-        run_crons(env)
+            c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=c1)
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            prx.post_comment('hansen delegate+', config['role_reviewer']['token'])
+            prx.post_comment('hansen r+', config['role_user']['token'])
+        env.run_crons()
 
         assert prx.user == users['user']
         assert env['runbot_merge.pull_requests'].search([
@@ -2243,41 +2375,44 @@ class TestReviewing(object):
             ('number', '=', prx.number)
         ]).state == 'ready'
 
-    def test_delegate_review_thirdparty(self, env, repo, users):
+    def test_delegate_review_thirdparty(self, env, repo, users, config):
         """Users should be able to delegate review to either the creator of
         the PR or an other user without review rights
         """
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
-        repo.make_ref('heads/master', m2)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
+            repo.make_ref('heads/master', m2)
 
-        c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c1, user='user')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        prx.post_comment('hansen delegate=%s' % users['other'], user='reviewer')
-        prx.post_comment('hansen r+', user='user')
+            c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=c1)
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            prx.post_comment('hansen delegate=%s' % users['other'], config['role_reviewer']['token'])
+            prx.post_comment('hansen r+', config['role_user']['token'])
+        env.run_crons()
 
-        run_crons(env)
         assert prx.user == users['user']
         assert env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
         ]).state == 'validated'
 
-        prx.post_comment('hansen r+', user='other')
+        with repo:
+            prx.post_comment('hansen r+', config['role_other']['token'])
         assert env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
         ]).state == 'ready'
 
-    def test_delegate_prefixes(self, env, repo):
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        repo.make_ref('heads/master', m)
+    def test_delegate_prefixes(self, env, repo, config):
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            repo.make_ref('heads/master', m)
 
-        c = repo.make_commit(m, 'first', None, tree={'m': 'c'})
-        prx = repo.make_pr('title', None, target='master', ctid=c, user='user')
-        prx.post_comment('hansen delegate=foo,@bar,#baz', user='reviewer')
+            c = repo.make_commit(m, 'first', None, tree={'m': 'c'})
+            prx = repo.make_pr(title='title', body=None, target='master', head=c)
+            prx.post_comment('hansen delegate=foo,@bar,#baz', config['role_reviewer']['token'])
 
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
@@ -2287,34 +2422,38 @@ class TestReviewing(object):
         assert {d.github_login for d in pr.delegates} == {'foo', 'bar', 'baz'}
 
 
-    def test_actual_review(self, env, repo):
+    def test_actual_review(self, env, repo, config):
         """ treat github reviews as regular comments
         """
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
-        repo.make_ref('heads/master', m2)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
+            repo.make_ref('heads/master', m2)
 
-        c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c1, user='user')
+            c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=c1)
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
         ])
 
-        prx.post_review('COMMENT', 'reviewer', "hansen priority=1")
+        with repo:
+            prx.post_review('COMMENT', "hansen priority=1", config['role_reviewer']['token'])
         assert pr.priority == 1
         assert pr.state == 'opened'
 
-        prx.post_review('APPROVE', 'reviewer', "hansen priority=2")
+        with repo:
+            prx.post_review('APPROVE', "hansen priority=2", config['role_reviewer']['token'])
         assert pr.priority == 2
         assert pr.state == 'opened'
 
-        prx.post_review('REQUEST_CHANGES', 'reviewer', 'hansen priority=1')
+        with repo:
+            prx.post_review('REQUEST_CHANGES', 'hansen priority=1', config['role_reviewer']['token'])
         assert pr.priority == 1
         assert pr.state == 'opened'
 
-
-        prx.post_review('COMMENT', 'reviewer', 'hansen r+')
+        with repo:
+            prx.post_review('COMMENT', 'hansen r+', config['role_reviewer']['token'])
         assert pr.priority == 1
         assert pr.state == 'approved'
 
@@ -2331,16 +2470,17 @@ class TestUnknownPR:
     * get statuses if head commit unknown (additional cron?)
     * handle any comment & review (existing PRs may enter the system on a review/r+)
     """
-    def test_rplus_unknown(self, repo, env):
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
-        repo.make_ref('heads/master', m2)
+    def test_rplus_unknown(self, repo, env, config):
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
+            repo.make_ref('heads/master', m2)
 
-        c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c1, user='user')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        repo.post_status(prx.head, 'success', 'ci/runbot', target_url="http://example.org/wheee")
-        run_crons(env)
+            c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=c1)
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            repo.post_status(prx.head, 'success', 'ci/runbot', target_url="http://example.org/wheee")
+        env.run_crons()
 
         # assume an unknown but ready PR: we don't know the PR or its head commit
         env['runbot_merge.pull_requests'].search([
@@ -2350,12 +2490,13 @@ class TestUnknownPR:
         env['runbot_merge.commit'].search([('sha', '=', prx.head)]).unlink()
 
         # reviewer reviewers
-        prx.post_comment('hansen r+', "reviewer")
+        with repo:
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
 
         Fetch = env['runbot_merge.fetch_job']
         assert Fetch.search([('repository', '=', repo.name), ('number', '=', prx.number)])
-        env['runbot_merge.project']._check_fetch()
-        run_crons(env)
+        env.run_crons('runbot_merge.fetch_prs_cron')
+        env.run_crons()
         assert not Fetch.search([('repository', '=', repo.name), ('number', '=', prx.number)])
 
         c = env['runbot_merge.commit'].search([('sha', '=', prx.head)])
@@ -2370,47 +2511,51 @@ class TestUnknownPR:
         ])
         assert pr.state == 'ready'
 
-        env['runbot_merge.project']._check_progress()
+        env.run_crons('runbot_merge.merge_cron')
         assert pr.staging_id
 
-    def test_rplus_unmanaged(self, env, repo, users):
+    def test_rplus_unmanaged(self, env, repo, users, config):
         """ r+ on an unmanaged target should notify about
         """
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
-        repo.make_ref('heads/branch', m2)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
+            repo.make_ref('heads/branch', m2)
 
-        c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='branch', ctid=c1, user='user')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
+            c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
+            prx = repo.make_pr(title='title', body='body', target='branch', head=c1)
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            repo.post_status(prx.head, 'success', 'ci/runbot')
 
-        prx.post_comment('hansen r+', "reviewer")
-
-        env['runbot_merge.project']._check_fetch()
-        env['runbot_merge.project']._send_feedback()
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
+        env.run_crons(
+            'runbot_merge.fetch_prs_cron',
+            'runbot_merge.feedback_cron',
+        )
 
         assert prx.comments == [
             (users['reviewer'], 'hansen r+'),
             (users['user'], "I'm sorry. Branch `branch` is not within my remit."),
         ]
 
-    def test_rplus_review_unmanaged(self, env, repo, users):
+    def test_rplus_review_unmanaged(self, env, repo, users, config):
         """ r+ reviews can take a different path than comments
         """
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
-        repo.make_ref('heads/branch', m2)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            m2 = repo.make_commit(m, 'second', None, tree={'m': 'm', 'm2': 'm2'})
+            repo.make_ref('heads/branch', m2)
 
-        c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='branch', ctid=c1, user='user')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
+            c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
+            prx = repo.make_pr(title='title', body='body', target='branch', head=c1)
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            repo.post_status(prx.head, 'success', 'ci/runbot')
 
-        prx.post_review('APPROVE', "reviewer", 'hansen r+')
-
-        env['runbot_merge.project']._check_fetch()
-        env['runbot_merge.project']._send_feedback()
+            prx.post_review('APPROVE', 'hansen r+', config['role_reviewer']['token'])
+        env.run_crons(
+            'runbot_merge.fetch_prs_cron',
+            'runbot_merge.feedback_cron',
+        )
 
         # FIXME: either split out reviews in local or merge reviews & comments in remote
         assert prx.comments[-1:] == [
@@ -2419,15 +2564,16 @@ class TestUnknownPR:
 
 class TestRecognizeCommands:
     @pytest.mark.parametrize('botname', ['hansen', 'Hansen', 'HANSEN', 'HanSen', 'hAnSeN'])
-    def test_botname_casing(self, repo, env, botname):
+    def test_botname_casing(self, repo, env, botname, config):
         """ Test that the botname is case-insensitive as people might write
         bot names capitalised or titlecased or uppercased or whatever
         """
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            repo.make_ref('heads/master', m)
 
-        c = repo.make_commit(m, 'first', None, tree={'m': 'c'})
-        prx = repo.make_pr('title', None, target='master', ctid=c, user='user')
+            c = repo.make_commit(m, 'first', None, tree={'m': 'c'})
+            prx = repo.make_pr(title='title', body=None, target='master', head=c)
 
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
@@ -2435,19 +2581,20 @@ class TestRecognizeCommands:
         ])
         assert pr.state == 'opened'
 
-        prx.post_comment('%s r+' % botname, 'reviewer')
+        with repo:
+            prx.post_comment('%s r+' % botname, config['role_reviewer']['token'])
         assert pr.state == 'approved'
 
     @pytest.mark.parametrize('indent', ['', '\N{SPACE}', '\N{SPACE}'*4, '\N{TAB}'])
-    def test_botname_indented(self, repo, env, indent):
+    def test_botname_indented(self, repo, env, indent, config):
         """ matching botname should ignore leading whitespaces
         """
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            repo.make_ref('heads/master', m)
 
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        repo.make_ref('heads/master', m)
-
-        c = repo.make_commit(m, 'first', None, tree={'m': 'c'})
-        prx = repo.make_pr('title', None, target='master', ctid=c, user='user')
+            c = repo.make_commit(m, 'first', None, tree={'m': 'c'})
+            prx = repo.make_pr(title='title', body=None, target='master', head=c)
 
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
@@ -2455,18 +2602,20 @@ class TestRecognizeCommands:
         ])
         assert pr.state == 'opened'
 
-        prx.post_comment('%shansen r+' % indent, 'reviewer')
+        with repo:
+            prx.post_comment('%shansen r+' % indent, config['role_reviewer']['token'])
         assert pr.state == 'approved'
 
 class TestRMinus:
-    def test_rminus_approved(self, repo, env):
+    def test_rminus_approved(self, repo, env, config):
         """ approved -> r- -> opened
         """
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            repo.make_ref('heads/master', m)
 
-        c = repo.make_commit(m, 'first', None, tree={'m': 'c'})
-        prx = repo.make_pr('title', None, target='master', ctid=c, user='user')
+            c = repo.make_commit(m, 'first', None, tree={'m': 'c'})
+            prx = repo.make_pr(title='title', body=None, target='master', head=c)
 
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
@@ -2474,31 +2623,37 @@ class TestRMinus:
         ])
         assert pr.state == 'opened'
 
-        prx.post_comment('hansen r+', 'reviewer')
+        with repo:
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
         assert pr.state == 'approved'
 
-        prx.post_comment('hansen r-', 'user')
+        with repo:
+            prx.post_comment('hansen r-', config['role_user']['token'])
         assert pr.state == 'opened'
-        prx.post_comment('hansen r+', 'reviewer')
+        with repo:
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
         assert pr.state == 'approved'
 
-        prx.post_comment('hansen r-', 'other')
+        with repo:
+            prx.post_comment('hansen r-', config['role_other']['token'])
         assert pr.state == 'approved'
 
-        prx.post_comment('hansen r-', 'reviewer')
+        with repo:
+            prx.post_comment('hansen r-', config['role_reviewer']['token'])
         assert pr.state == 'opened'
 
-    def test_rminus_ready(self, repo, env):
+    def test_rminus_ready(self, repo, env, config):
         """ ready -> r- -> validated
         """
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            repo.make_ref('heads/master', m)
 
-        c = repo.make_commit(m, 'first', None, tree={'m': 'c'})
-        prx = repo.make_pr('title', None, target='master', ctid=c, user='user')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        run_crons(env)
+            c = repo.make_commit(m, 'first', None, tree={'m': 'c'})
+            prx = repo.make_pr(title='title', body=None, target='master', head=c)
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            repo.post_status(prx.head, 'success', 'legal/cla')
+        env.run_crons()
 
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
@@ -2506,31 +2661,37 @@ class TestRMinus:
         ])
         assert pr.state == 'validated'
 
-        prx.post_comment('hansen r+', 'reviewer')
+        with repo:
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
         assert pr.state == 'ready'
 
-        prx.post_comment('hansen r-', 'user')
+        with repo:
+            prx.post_comment('hansen r-', config['role_user']['token'])
         assert pr.state == 'validated'
-        prx.post_comment('hansen r+', 'reviewer')
+        with repo:
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
         assert pr.state == 'ready'
 
-        prx.post_comment('hansen r-', 'other')
+        with repo:
+            prx.post_comment('hansen r-', config['role_other']['token'])
         assert pr.state == 'ready'
 
-        prx.post_comment('hansen r-', 'reviewer')
+        with repo:
+            prx.post_comment('hansen r-', config['role_reviewer']['token'])
         assert pr.state == 'validated'
 
-    def test_rminus_staged(self, repo, env):
+    def test_rminus_staged(self, repo, env, config):
         """ staged -> r- -> validated
         """
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            repo.make_ref('heads/master', m)
 
-        c = repo.make_commit(m, 'first', None, tree={'m': 'c'})
-        prx = repo.make_pr('title', None, target='master', ctid=c, user='user')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        run_crons(env)
+            c = repo.make_commit(m, 'first', None, tree={'m': 'c'})
+            prx = repo.make_pr(title='title', body=None, target='master', head=c)
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            repo.post_status(prx.head, 'success', 'legal/cla')
+        env.run_crons()
 
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
@@ -2538,57 +2699,65 @@ class TestRMinus:
         ])
 
         # if reviewer unreviews, cancel staging & unreview
-        prx.post_comment('hansen r+', 'reviewer')
-        run_crons(env)
+        with repo:
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
+        env.run_crons()
         st = pr.staging_id
         assert st
 
-        prx.post_comment('hansen r-', 'reviewer')
+        with repo:
+            prx.post_comment('hansen r-', config['role_reviewer']['token'])
         assert not st.active
         assert not pr.staging_id
         assert pr.state == 'validated'
 
         # if author unreviews, cancel staging & unreview
-        prx.post_comment('hansen r+', 'reviewer')
-        run_crons(env)
+        with repo:
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
+        env.run_crons()
         st = pr.staging_id
         assert st
 
-        prx.post_comment('hansen r-', 'user')
+        with repo:
+            prx.post_comment('hansen r-', config['role_user']['token'])
         assert not st.active
         assert not pr.staging_id
         assert pr.state == 'validated'
 
         # if rando unreviews, ignore
-        prx.post_comment('hansen r+', 'reviewer')
-        run_crons(env)
+        with repo:
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
+        env.run_crons()
         st = pr.staging_id
         assert st
 
-        prx.post_comment('hansen r-', 'other')
+        with repo:
+            prx.post_comment('hansen r-', config['role_other']['token'])
         assert pr.staging_id == st
         assert pr.state == 'ready'
 
-    def test_split(self, env, repo):
+    def test_split(self, env, repo, config):
         """ Should remove the PR from its split, and possibly delete the split
         entirely.
         """
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            repo.make_ref('heads/master', m)
 
-        c = repo.make_commit(m, 'first', None, tree={'m': 'm', '1': '1'})
-        prx1 = repo.make_pr('t1', 'b1', target='master', ctid=c, user='user', label='p1')
-        repo.post_status(prx1.head, 'success', 'legal/cla')
-        repo.post_status(prx1.head, 'success', 'ci/runbot')
-        prx1.post_comment('hansen r+', user='reviewer')
+            c = repo.make_commit(m, 'first', None, tree={'m': 'm', '1': '1'})
+            repo.make_ref('heads/p1', c)
+            prx1 = repo.make_pr(title='t1', body='b1', target='master', head='p1')
+            repo.post_status(prx1.head, 'success', 'legal/cla')
+            repo.post_status(prx1.head, 'success', 'ci/runbot')
+            prx1.post_comment('hansen r+', config['role_reviewer']['token'])
 
-        c = repo.make_commit(m, 'first', None, tree={'m': 'm', '2': '2'})
-        prx2 = repo.make_pr('t2', 'b2', target='master', ctid=c, user='user', label='p2')
-        repo.post_status(prx2.head, 'success', 'legal/cla')
-        repo.post_status(prx2.head, 'success', 'ci/runbot')
-        prx2.post_comment('hansen r+', user='reviewer')
-
-        run_crons(env)
+            c = repo.make_commit(m, 'first', None, tree={'m': 'm', '2': '2'})
+            repo.make_ref('heads/p2', c)
+            prx2 = repo.make_pr(title='t2', body='b2', target='master', head='p2')
+            repo.post_status(prx2.head, 'success', 'legal/cla')
+            repo.post_status(prx2.head, 'success', 'ci/runbot')
+            prx2.post_comment('hansen r+', config['role_reviewer']['token'])
+        env.run_crons()
 
         pr1, pr2 = env['runbot_merge.pull_requests'].search([], order='number')
         assert pr1.number == prx1.number
@@ -2596,8 +2765,9 @@ class TestRMinus:
         assert pr1.staging_id == pr2.staging_id
         s0 = pr1.staging_id
 
-        repo.post_status('heads/staging.master', 'failure', 'ci/runbot')
-        run_crons(env)
+        with repo:
+            repo.post_status('heads/staging.master', 'failure', 'ci/runbot')
+        env.run_crons()
 
         assert pr1.staging_id and pr1.staging_id != s0, "pr1 should have been re-staged"
         assert not pr2.staging_id, "pr2 should not"
@@ -2605,27 +2775,29 @@ class TestRMinus:
         #       around using active_test=False
         assert env['runbot_merge.split'].search([])
 
-        # prx2 was actually a terrible idea!
-        prx2.post_comment('hansen r-', user='reviewer')
+        with repo:
+            # prx2 was actually a terrible idea!
+            prx2.post_comment('hansen r-', config['role_reviewer']['token'])
         # probably not necessary ATM but...
-        run_crons(env)
+        env.run_crons()
 
         assert pr2.state == 'validated', "state should have been reset"
         assert not env['runbot_merge.split'].search([]), "there should be no split left"
 
 class TestComments:
-    def test_address_method(self, repo, env):
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        repo.make_ref('heads/master', m)
+    def test_address_method(self, repo, env, config):
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            repo.make_ref('heads/master', m)
 
-        c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c1, user='user')
+            c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=c1)
 
-        repo.post_status(prx.head, 'success', 'legal/cla')
-        repo.post_status(prx.head, 'success', 'ci/runbot')
-        prx.post_comment('hansen delegate=foo', user='reviewer')
-        prx.post_comment('@hansen delegate=bar', user='reviewer')
-        prx.post_comment('#hansen delegate=baz', user='reviewer')
+            repo.post_status(prx.head, 'success', 'legal/cla')
+            repo.post_status(prx.head, 'success', 'ci/runbot')
+            prx.post_comment('hansen delegate=foo', config['role_reviewer']['token'])
+            prx.post_comment('@hansen delegate=bar', config['role_reviewer']['token'])
+            prx.post_comment('#hansen delegate=baz', config['role_reviewer']['token'])
 
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
@@ -2635,92 +2807,104 @@ class TestComments:
         assert {p.github_login for p in pr.delegates} \
             == {'foo', 'bar', 'baz'}
 
-    def test_delete(self, repo, env):
+    def test_delete(self, repo, env, config):
         """ Comments being deleted should be ignored
         """
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            repo.make_ref('heads/master', m)
 
-        c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c1, user='user')
+            c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=c1)
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
         ])
 
-        cid = prx.post_comment('hansen r+', user='reviewer')
-        # unreview by pushing a new commit
-        prx.push(repo.make_commit(c1, 'second', None, tree={'m': 'c2'}))
+        with repo:
+            cid = prx.post_comment('hansen r+', config['role_reviewer']['token'])
+            # unreview by pushing a new commit
+            repo.update_ref(prx.ref, repo.make_commit(c1, 'second', None, tree={'m': 'c2'}), force=True)
         assert pr.state == 'opened'
-        prx.delete_comment(cid, 'reviewer')
+        with repo:
+            prx.delete_comment(cid, config['role_reviewer']['token'])
         # check that PR is still unreviewed
         assert pr.state == 'opened'
 
-    def test_edit(self, repo, env):
+    def test_edit(self, repo, env, config):
         """ Comments being edited should be ignored
         """
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            repo.make_ref('heads/master', m)
 
-        c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c1, user='user')
+            c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=c1)
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
         ])
 
-        cid = prx.post_comment('hansen r+', user='reviewer')
-        # unreview by pushing a new commit
-        prx.push(repo.make_commit(c1, 'second', None, tree={'m': 'c2'}))
+        with repo:
+            cid = prx.post_comment('hansen r+', config['role_reviewer']['token'])
+            # unreview by pushing a new commit
+            repo.update_ref(prx.ref, repo.make_commit(c1, 'second', None, tree={'m': 'c2'}), force=True)
         assert pr.state == 'opened'
-        prx.edit_comment(cid, 'hansen r+ edited', 'reviewer')
+        with repo:
+            prx.edit_comment(cid, 'hansen r+ edited', config['role_reviewer']['token'])
         # check that PR is still unreviewed
         assert pr.state == 'opened'
 
 class TestFeedback:
-    def test_ci_approved(self, repo, env, users):
+    def test_ci_approved(self, repo, env, users, config):
         """CI failing on an r+'d PR sends feedback"""
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            repo.make_ref('heads/master', m)
 
-        c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c1, user='user')
+            c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=c1)
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
         ])
 
-        prx.post_comment('hansen r+', user='reviewer')
+        with repo:
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
         assert pr.state == 'approved'
 
-        repo.post_status(prx.head, 'failure', 'ci/runbot')
-        run_crons(env)
+        with repo:
+            repo.post_status(prx.head, 'failure', 'ci/runbot')
+        env.run_crons()
 
         assert prx.comments == [
             (users['reviewer'], 'hansen r+'),
             (users['user'], "'ci/runbot' failed on this reviewed PR.")
         ]
 
-    def test_review_unvalidated(self, repo, env, users):
+    def test_review_unvalidated(self, repo, env, users, config):
         """r+-ing a PR with failed CI sends feedback"""
-        m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
-        repo.make_ref('heads/master', m)
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'m': 'm'})
+            repo.make_ref('heads/master', m)
 
-        c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
-        prx = repo.make_pr('title', 'body', target='master', ctid=c1, user='user')
+            c1 = repo.make_commit(m, 'first', None, tree={'m': 'c1'})
+            prx = repo.make_pr(title='title', body='body', target='master', head=c1)
         pr = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', prx.number)
         ])
 
-        repo.post_status(prx.head, 'failure', 'ci/runbot')
-        run_crons(env)
+        with repo:
+            repo.post_status(prx.head, 'failure', 'ci/runbot')
+        env.run_crons()
         assert pr.state == 'opened'
 
-        prx.post_comment('hansen r+', user='reviewer')
+        with repo:
+            prx.post_comment('hansen r+', config['role_reviewer']['token'])
         assert pr.state == 'approved'
 
-        run_crons(env)
+        env.run_crons()
 
         assert prx.comments == [
             (users['reviewer'], 'hansen r+'),
@@ -2730,14 +2914,15 @@ class TestInfrastructure:
     def test_protection(self, repo):
         """ force-pushing on a protected ref should fail
         """
-        m0 = repo.make_commit(None, 'initial', None, tree={'m': 'm0'})
-        m1 = repo.make_commit(m0, 'first', None, tree={'m': 'm1'})
-        repo.make_ref('heads/master', m1)
-        repo.protect('master')
+        with repo:
+            m0 = repo.make_commit(None, 'initial', None, tree={'m': 'm0'})
+            m1 = repo.make_commit(m0, 'first', None, tree={'m': 'm1'})
+            repo.make_ref('heads/master', m1)
+            repo.protect('master')
 
-        c1 = repo.make_commit(m0, 'other', None, tree={'m': 'c1'})
-        with pytest.raises(AssertionError):
-            repo.update_ref('heads/master', c1, force=True)
+            c1 = repo.make_commit(m0, 'other', None, tree={'m': 'c1'})
+            with pytest.raises(AssertionError):
+                repo.update_ref('heads/master', c1, force=True)
         assert repo.get_ref('heads/master') == m1
 
 def node(name, *children):
@@ -2782,55 +2967,61 @@ class TestEmailFormatting:
         assert p1.formatted_email == 'Shultz <Osmose99@users.noreply.github.com>'
 
 class TestLabelling:
-    def test_desync(self, env, repo):
-        m = repo.make_commit(None, 'initial', None, tree={'a': 'some content'})
-        repo.make_ref('heads/master', m)
+    def test_desync(self, env, repo, config):
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'a': 'some content'})
+            repo.make_ref('heads/master', m)
 
-        c = repo.make_commit(m, 'replace file contents', None, tree={'a': 'some other content'})
-        pr = repo.make_pr('gibberish', 'blahblah', target='master', ctid=c, user='user')
+            c = repo.make_commit(m, 'replace file contents', None, tree={'a': 'some other content'})
+            pr = repo.make_pr(title='gibberish', body='blahblah', target='master', head=c)
 
         [pr_id] = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', pr.number),
         ])
-        repo.post_status(c, 'success', 'legal/cla')
-        repo.post_status(c, 'success', 'ci/runbot')
+        with repo:
+            repo.post_status(c, 'success', 'legal/cla')
+            repo.post_status(c, 'success', 'ci/runbot')
 
-        run_crons(env)
+        env.run_crons()
 
         assert pr.labels == {'seen 🙂', 'CI 🤖'}
-        # desync state and labels
-        pr.labels.remove('CI 🤖')
+        with repo:
+            # desync state and labels
+            pr.labels.remove('CI 🤖')
 
-        pr.post_comment('hansen r+', 'reviewer')
-        run_crons(env)
+            pr.post_comment('hansen r+', config['role_reviewer']['token'])
+        env.run_crons()
 
         assert pr.labels == {'seen 🙂', 'CI 🤖', 'r+ 👌', 'merging 👷'},\
             "labels should be resynchronised"
 
-    def test_other_tags(self, env, repo):
-        m = repo.make_commit(None, 'initial', None, tree={'a': 'some content'})
-        repo.make_ref('heads/master', m)
+    def test_other_tags(self, env, repo, config):
+        with repo:
+            m = repo.make_commit(None, 'initial', None, tree={'a': 'some content'})
+            repo.make_ref('heads/master', m)
 
-        c = repo.make_commit(m, 'replace file contents', None, tree={'a': 'some other content'})
-        pr = repo.make_pr('gibberish', 'blahblah', target='master', ctid=c, user='user')
+            c = repo.make_commit(m, 'replace file contents', None, tree={'a': 'some other content'})
+            pr = repo.make_pr(title='gibberish', body='blahblah', target='master', head=c)
 
-        # "foreign" labels
-        pr.labels.update(('L1', 'L2'))
+        with repo:
+            # "foreign" labels
+            pr.labels.update(('L1', 'L2'))
 
         [pr_id] = env['runbot_merge.pull_requests'].search([
             ('repository.name', '=', repo.name),
             ('number', '=', pr.number),
         ])
-        repo.post_status(c, 'success', 'legal/cla')
-        repo.post_status(c, 'success', 'ci/runbot')
-
-        run_crons(env)
+        with repo:
+            repo.post_status(c, 'success', 'legal/cla')
+            repo.post_status(c, 'success', 'ci/runbot')
+        env.run_crons()
 
         assert pr.labels == {'seen 🙂', 'CI 🤖', 'L1', 'L2'}, "should not lose foreign labels"
 
-        pr.post_comment('hansen r+', 'reviewer')
-        run_crons(env)
+        with repo:
+            pr.post_comment('hansen r+', config['role_reviewer']['token'])
+        env.run_crons()
 
         assert pr.labels == {'seen 🙂', 'CI 🤖', 'r+ 👌', 'merging 👷', 'L1', 'L2'},\
             "should not lose foreign labels"
