+++ runbot_merge/tests/test_multirepo.py
@@ -9,23 +9,29 @@ import json
 
 import pytest
 
-from test_utils import re_matches, run_crons, get_partner
+from test_utils import re_matches, get_partner
 
 @pytest.fixture
-def repo_a(make_repo):
-    return make_repo('a')
+def repo_a(project, make_repo):
+    repo = make_repo('a')
+    project.write({'repo_ids': [(0, 0, {'name': repo.name})]})
+    return repo
 
 @pytest.fixture
-def repo_b(make_repo):
-    return make_repo('b')
+def repo_b(project, make_repo):
+    repo = make_repo('b')
+    project.write({'repo_ids': [(0, 0, {'name': repo.name})]})
+    return repo
 
 @pytest.fixture
-def repo_c(make_repo):
-    return make_repo('c')
+def repo_c(project, make_repo):
+    repo = make_repo('c')
+    project.write({'repo_ids': [(0, 0, {'name': repo.name})]})
+    return repo
 
-def make_pr(repo, prefix, trees, *, target='master', user='user', label=None,
+def make_pr(repo, prefix, trees, *, target='master', user,
             statuses=(('ci/runbot', 'success'), ('legal/cla', 'success')),
-            reviewer='reviewer'):
+            reviewer):
     """
     :type repo: fake_github.Repo
     :type prefix: str
@@ -37,14 +43,16 @@ def make_pr(repo, prefix, trees, *, target='master', user='user', label=None,
     :type reviewer: str | None
     :rtype: fake_github.PR
     """
-    base = repo.commit('heads/{}'.format(target))
-    tree = repo.read_tree(base)
-    c = base.id
-    for i, t in enumerate(trees):
-        tree.update(t)
-        c = repo.make_commit(c, 'commit_{}_{:02}'.format(prefix, i), None,
-                             tree=dict(tree))
-    pr = repo.make_pr('title {}'.format(prefix), 'body {}'.format(prefix), target=target, ctid=c, user=user, label=label)
+    *_, c = repo.make_commits(
+        'heads/{}'.format(target),
+        *(
+            repo.Commit('commit_{}_{:02}'.format(prefix, i), tree=tree)
+            for i, tree in enumerate(trees)
+        ),
+        ref='heads/{}'.format(prefix)
+    )
+    pr = repo.make_pr(title='title {}'.format(prefix), body='body {}'.format(prefix),
+                      target=target, head=prefix, token=user)
     for context, result in statuses:
         repo.post_status(c, result, context)
     if reviewer:
@@ -62,35 +70,51 @@ def make_branch(repo, name, message, tree, protect=True):
         repo.protect(name)
     return c
 
-def test_stage_one(env, project, repo_a, repo_b):
+def test_stage_one(env, project, repo_a, repo_b, config):
     """ First PR is non-matched from A => should not select PR from B
     """
     project.batch_limit = 1
 
-    make_branch(repo_a, 'master', 'initial', {'a': 'a_0'})
-    pr_a = make_pr(repo_a, 'A', [{'a': 'a_1'}], label='do-a-thing')
+    with repo_a:
+        make_branch(repo_a, 'master', 'initial', {'a': 'a_0'})
+        pr_a = make_pr(
+            repo_a, 'A', [{'a': 'a_1'}],
+            user=config['role_user']['token'],
+            reviewer=config['role_reviewer']['token'])
 
-    make_branch(repo_b, 'master', 'initial', {'a': 'b_0'})
-    pr_b = make_pr(repo_b, 'B', [{'a': 'b_1'}], label='do-other-thing')
-
-    run_crons(env)
+    with repo_b:
+        make_branch(repo_b, 'master', 'initial', {'a': 'b_0'})
+        pr_b = make_pr(
+            repo_b, 'B', [{'a': 'b_1'}],
+            user=config['role_user']['token'],
+            reviewer=config['role_reviewer']['token'],
+        )
+    env.run_crons()
 
     assert to_pr(env, pr_a).state == 'ready'
     assert to_pr(env, pr_a).staging_id
     assert to_pr(env, pr_b).state == 'ready'
     assert not to_pr(env, pr_b).staging_id
 
-def test_stage_match(env, project, repo_a, repo_b):
+def test_stage_match(env, project, repo_a, repo_b, config):
     """ First PR is matched from A,  => should select matched PR from B
     """
     project.batch_limit = 1
-    make_branch(repo_a, 'master', 'initial', {'a': 'a_0'})
-    pr_a = make_pr(repo_a, 'A', [{'a': 'a_1'}], label='do-a-thing')
 
-    make_branch(repo_b, 'master', 'initial', {'a': 'b_0'})
-    pr_b = make_pr(repo_b, 'B', [{'a': 'b_1'}], label='do-a-thing')
-
-    run_crons(env)
+    with repo_a:
+        make_branch(repo_a, 'master', 'initial', {'a': 'a_0'})
+        pr_a = make_pr(
+            repo_a, 'do-a-thing', [{'a': 'a_1'}],
+            user=config['role_user']['token'],
+            reviewer=config['role_reviewer']['token'],
+        )
+    with repo_b:
+        make_branch(repo_b, 'master', 'initial', {'a': 'b_0'})
+        pr_b = make_pr(repo_b, 'do-a-thing', [{'a': 'b_1'}],
+            user=config['role_user']['token'],
+            reviewer=config['role_reviewer']['token'],
+        )
+    env.run_crons()
 
     pr_a = to_pr(env, pr_a)
     pr_b = to_pr(env, pr_b)
@@ -102,7 +126,7 @@ def test_stage_match(env, project, repo_a, repo_b):
     assert pr_a.staging_id == pr_b.staging_id, \
         "branch-matched PRs should be part of the same staging"
 
-def test_unmatch_patch(env, project, repo_a, repo_b):
+def test_unmatch_patch(env, project, repo_a, repo_b, config):
     """ When editing files via the UI for a project you don't have write
     access to, a branch called patch-XXX is automatically created in your
     profile to hold the change.
@@ -115,13 +139,21 @@ def test_unmatch_patch(env, project, repo_a, repo_b):
     -> PRs with a branch name of patch-* should not be label-matched
     """
     project.batch_limit = 1
-    make_branch(repo_a, 'master', 'initial', {'a': 'a_0'})
-    pr_a = make_pr(repo_a, 'A', [{'a': 'a_1'}], label='patch-1')
-
-    make_branch(repo_b, 'master', 'initial', {'a': 'b_0'})
-    pr_b = make_pr(repo_b, 'B', [{'a': 'b_1'}], label='patch-1')
-
-    run_crons(env)
+    with repo_a:
+        make_branch(repo_a, 'master', 'initial', {'a': 'a_0'})
+        pr_a = make_pr(
+            repo_a, 'patch-1', [{'a': 'a_1'}],
+            user=config['role_user']['token'],
+            reviewer=config['role_reviewer']['token'],
+        )
+    with repo_b:
+        make_branch(repo_b, 'master', 'initial', {'a': 'b_0'})
+        pr_b = make_pr(
+            repo_b, 'patch-1', [{'a': 'b_1'}],
+            user=config['role_user']['token'],
+            reviewer=config['role_reviewer']['token'],
+        )
+    env.run_crons()
 
     pr_a = to_pr(env, pr_a)
     pr_b = to_pr(env, pr_b)
@@ -130,20 +162,27 @@ def test_unmatch_patch(env, project, repo_a, repo_b):
     assert pr_b.state == 'ready'
     assert not pr_b.staging_id, 'patch-* PRs should not be branch-matched'
 
-def test_sub_match(env, project, repo_a, repo_b, repo_c):
+def test_sub_match(env, project, repo_a, repo_b, repo_c, config):
     """ Branch-matching should work on a subset of repositories
     """
     project.batch_limit = 1
-    make_branch(repo_a, 'master', 'initial', {'a': 'a_0'})
-    # no pr here
-
-    make_branch(repo_b, 'master', 'initial', {'a': 'b_0'})
-    pr_b = make_pr(repo_b, 'B', [{'a': 'b_1'}], label='do-a-thing')
-
-    make_branch(repo_c, 'master', 'initial', {'a': 'c_0'})
-    pr_c = make_pr(repo_c, 'C', [{'a': 'c_1'}], label='do-a-thing')
-
-    run_crons(env)
+    with repo_a: # no pr here
+        make_branch(repo_a, 'master', 'initial', {'a': 'a_0'})
+    with repo_b:
+        make_branch(repo_b, 'master', 'initial', {'a': 'b_0'})
+        pr_b = make_pr(
+            repo_b, 'do-a-thing', [{'a': 'b_1'}],
+            user=config['role_user']['token'],
+            reviewer=config['role_reviewer']['token'],
+        )
+    with repo_c:
+        make_branch(repo_c, 'master', 'initial', {'a': 'c_0'})
+        pr_c = make_pr(
+            repo_c, 'do-a-thing', [{'a': 'c_1'}],
+            user=config['role_user']['token'],
+            reviewer=config['role_reviewer']['token'],
+        )
+    env.run_crons()
 
     pr_b = to_pr(env, pr_b)
     pr_c = to_pr(env, pr_c)
@@ -167,28 +206,44 @@ def test_sub_match(env, project, repo_a, repo_b, repo_c):
         repo_c.name + '^': c_staging.parents[0],
     }
 
-def test_merge_fail(env, project, repo_a, repo_b, users):
+def test_merge_fail(env, project, repo_a, repo_b, users, config):
     """ In a matched-branch scenario, if merging in one of the linked repos
     fails it should revert the corresponding merges
     """
     project.batch_limit = 1
 
-    make_branch(repo_a, 'master', 'initial', {'a': 'a_0'})
-    make_branch(repo_b, 'master', 'initial', {'a': 'b_0'})
+    with repo_a, repo_b:
+        make_branch(repo_a, 'master', 'initial', {'a': 'a_0'})
+        make_branch(repo_b, 'master', 'initial', {'a': 'b_0'})
 
-    # first set of matched PRs
-    pr1a = make_pr(repo_a, 'A', [{'a': 'a_1'}], label='do-a-thing')
-    pr1b = make_pr(repo_b, 'B', [{'a': 'b_1'}], label='do-a-thing')
+        # first set of matched PRs
+        pr1a = make_pr(
+            repo_a, 'do-a-thing', [{'a': 'a_1'}],
+            user=config['role_user']['token'],
+            reviewer=config['role_reviewer']['token'],
+        )
+        pr1b = make_pr(
+            repo_b, 'do-a-thing', [{'a': 'b_1'}],
+            user=config['role_user']['token'],
+            reviewer=config['role_reviewer']['token'],
+        )
 
-    # add a conflicting commit to B so the staging fails
-    repo_b.make_commit('heads/master', 'cn', None, tree={'a': 'cn'})
+        # add a conflicting commit to B so the staging fails
+        repo_b.make_commit('heads/master', 'cn', None, tree={'a': 'cn'})
 
-    # and a second set of PRs which should get staged while the first set
-    # fails
-    pr2a = make_pr(repo_a, 'A2', [{'b': 'ok'}], label='do-b-thing')
-    pr2b = make_pr(repo_b, 'B2', [{'b': 'ok'}], label='do-b-thing')
-
-    run_crons(env)
+        # and a second set of PRs which should get staged while the first set
+        # fails
+        pr2a = make_pr(
+            repo_a, 'do-b-thing', [{'b': 'ok'}],
+            user=config['role_user']['token'],
+            reviewer=config['role_reviewer']['token'],
+        )
+        pr2b = make_pr(
+            repo_b, 'do-b-thing', [{'b': 'ok'}],
+            user=config['role_user']['token'],
+            reviewer=config['role_reviewer']['token'],
+        )
+    env.run_crons()
 
     s2 = to_pr(env, pr2a) | to_pr(env, pr2b)
     st = env['runbot_merge.stagings'].search([])
@@ -208,33 +263,43 @@ def test_merge_fail(env, project, repo_a, repo_b, users):
         for c in repo_a.log('heads/staging.master')
     ] == [
         re_matches('^force rebuild'),
-        'commit_A2_00\n\ncloses %s#2\n\nSigned-off-by: %s' % (repo_a.name, reviewer),
+        'commit_do-b-thing_00\n\ncloses %s#2\n\nSigned-off-by: %s' % (repo_a.name, reviewer),
         'initial'
     ], "dummy commit + squash-merged PR commit + root commit"
 
-def test_ff_fail(env, project, repo_a, repo_b):
+def test_ff_fail(env, project, repo_a, repo_b, config):
     """ In a matched-branch scenario, fast-forwarding one of the repos fails
     the entire thing should be rolled back
     """
     project.batch_limit = 1
-    root_a = make_branch(repo_a, 'master', 'initial', {'a': 'a_0'})
-    make_pr(repo_a, 'A', [{'a': 'a_1'}], label='do-a-thing')
 
-    make_branch(repo_b, 'master', 'initial', {'a': 'b_0'})
-    make_pr(repo_b, 'B', [{'a': 'b_1'}], label='do-a-thing')
+    with repo_a, repo_b:
+        root_a = make_branch(repo_a, 'master', 'initial', {'a': 'a_0'})
+        make_pr(
+            repo_a, 'do-a-thing', [{'a': 'a_1'}],
+            user=config['role_user']['token'],
+            reviewer=config['role_reviewer']['token'],
+        )
 
-    run_crons(env)
+        make_branch(repo_b, 'master', 'initial', {'a': 'b_0'})
+        make_pr(
+            repo_b, 'do-a-thing', [{'a': 'b_1'}],
+            user=config['role_user']['token'],
+            reviewer=config['role_reviewer']['token'],
+        )
+    env.run_crons()
 
     # add second commit blocking FF
-    cn = repo_b.make_commit('heads/master', 'second', None, tree={'a': 'b_0', 'b': 'other'})
+    with repo_b:
+        cn = repo_b.make_commit('heads/master', 'second', None, tree={'a': 'b_0', 'b': 'other'})
     assert repo_b.commit('heads/master').id == cn
 
-    repo_a.post_status('heads/staging.master', 'success', 'ci/runbot')
-    repo_a.post_status('heads/staging.master', 'success', 'legal/cla')
-    repo_b.post_status('heads/staging.master', 'success', 'ci/runbot')
-    repo_b.post_status('heads/staging.master', 'success', 'legal/cla')
-
-    env['runbot_merge.project']._check_progress()
+    with repo_a, repo_b:
+        repo_a.post_status('heads/staging.master', 'success', 'ci/runbot')
+        repo_a.post_status('heads/staging.master', 'success', 'legal/cla')
+        repo_b.post_status('heads/staging.master', 'success', 'ci/runbot')
+        repo_b.post_status('heads/staging.master', 'success', 'legal/cla')
+    env.run_crons('runbot_merge.merge_cron')
     assert repo_b.commit('heads/master').id == cn,\
         "B should still be at the conflicting commit"
     assert repo_a.commit('heads/master').id == root_a,\
@@ -246,23 +311,32 @@ def test_ff_fail(env, project, repo_a, repo_b):
     assert len(st.batch_ids.prs) == 2
 
 class TestCompanionsNotReady:
-    def test_one_pair(self, env, project, repo_a, repo_b, owner, users):
+    def test_one_pair(self, env, project, repo_a, repo_b, config, users):
         """ If the companion of a ready branch-matched PR is not ready,
         they should not get staged
         """
         project.batch_limit = 1
-        make_branch(repo_a, 'master', 'initial', {'a': 'a_0'})
-        # pr_a is born ready
-        p_a = make_pr(repo_a, 'A', [{'a': 'a_1'}], label='do-a-thing')
+        with repo_a, repo_b:
+            make_branch(repo_a, 'master', 'initial', {'a': 'a_0'})
+            # pr_a is born ready
+            p_a = make_pr(
+                repo_a, 'do-a-thing', [{'a': 'a_1'}],
+                user=config['role_user']['token'],
+                reviewer=config['role_reviewer']['token'],
+            )
 
-        make_branch(repo_b, 'master', 'initial', {'a': 'b_0'})
-        p_b = make_pr(repo_b, 'B', [{'a': 'b_1'}], label='do-a-thing', reviewer=None)
+            make_branch(repo_b, 'master', 'initial', {'a': 'b_0'})
+            p_b = make_pr(
+                repo_b, 'do-a-thing', [{'a': 'b_1'}],
+                user=config['role_user']['token'],
+                reviewer=None,
+            )
 
         pr_a = to_pr(env, p_a)
         pr_b = to_pr(env, p_b)
-        assert pr_a.label == pr_b.label == '{}:do-a-thing'.format(owner)
+        assert pr_a.label == pr_b.label == '{}:do-a-thing'.format(config['github']['owner'])
 
-        run_crons(env)
+        env.run_crons()
 
         assert pr_a.state == 'ready'
         assert pr_b.state == 'validated'
@@ -275,28 +349,41 @@ class TestCompanionsNotReady:
             (users['user'], "Linked pull request(s) %s#%d not ready. Linked PRs are not staged until all of them are ready." % (repo_b.name, p_b.number)),
         ]
         # ensure the message is only sent once per PR
-        env['runbot_merge.pull_requests']._check_linked_prs_statuses()
+        env.run_crons('runbot_merge.check_linked_prs_status')
         assert p_a.comments == [
             (users['reviewer'], 'hansen r+'),
             (users['user'], "Linked pull request(s) %s#%d not ready. Linked PRs are not staged until all of them are ready." % (repo_b.name, p_b.number)),
         ]
         assert p_b.comments == []
 
-    def test_two_of_three_unready(self, env, project, repo_a, repo_b, repo_c, owner, users):
+    def test_two_of_three_unready(self, env, project, repo_a, repo_b, repo_c, users, config):
         """ In a 3-batch, if two of the PRs are not ready both should be
         linked by the first one
         """
         project.batch_limit = 1
-        make_branch(repo_a, 'master', 'initial', {'f': 'a0'})
-        pr_a = make_pr(repo_a, 'A', [{'f': 'a1'}], label='a-thing', reviewer=None)
+        with repo_a, repo_b, repo_c:
+            make_branch(repo_a, 'master', 'initial', {'f': 'a0'})
+            pr_a = make_pr(
+                repo_a, 'a-thing', [{'f': 'a1'}],
+                user=config['role_user']['token'],
+                reviewer=None,
+            )
 
-        make_branch(repo_b, 'master', 'initial', {'f': 'b0'})
-        pr_b = make_pr(repo_b, 'B', [{'f': 'b1'}], label='a-thing')
+            make_branch(repo_b, 'master', 'initial', {'f': 'b0'})
+            pr_b = make_pr(
+                repo_b, 'a-thing', [{'f': 'b1'}],
+                user=config['role_user']['token'],
+                reviewer=config['role_reviewer']['token'],
+            )
 
-        make_branch(repo_c, 'master', 'initial', {'f': 'c0'})
-        pr_c = make_pr(repo_c, 'C', [{'f': 'c1'}], label='a-thing', reviewer=None)
+            make_branch(repo_c, 'master', 'initial', {'f': 'c0'})
+            pr_c = make_pr(
+                repo_c, 'a-thing', [{'f': 'c1'}],
+                user=config['role_user']['token'],
+                reviewer=None,
+            )
+        env.run_crons()
 
-        run_crons(env)
         assert pr_a.comments == []
         assert pr_b.comments == [
             (users['reviewer'], 'hansen r+'),
@@ -307,21 +394,34 @@ class TestCompanionsNotReady:
         ]
         assert pr_c.comments == []
 
-    def test_one_of_three_unready(self, env, project, repo_a, repo_b, repo_c, owner, users):
+    def test_one_of_three_unready(self, env, project, repo_a, repo_b, repo_c, users, config):
         """ In a 3-batch, if one PR is not ready it should be linked on the
         other two
         """
         project.batch_limit = 1
-        make_branch(repo_a, 'master', 'initial', {'f': 'a0'})
-        pr_a = make_pr(repo_a, 'A', [{'f': 'a1'}], label='a-thing', reviewer=None)
+        with repo_a, repo_b, repo_c:
+            make_branch(repo_a, 'master', 'initial', {'f': 'a0'})
+            pr_a = make_pr(
+                repo_a, 'a-thing', [{'f': 'a1'}],
+                user=config['role_user']['token'],
+                reviewer=None,
+            )
 
-        make_branch(repo_b, 'master', 'initial', {'f': 'b0'})
-        pr_b = make_pr(repo_b, 'B', [{'f': 'b1'}], label='a-thing')
+            make_branch(repo_b, 'master', 'initial', {'f': 'b0'})
+            pr_b = make_pr(
+                repo_b, 'a-thing', [{'f': 'b1'}],
+                user=config['role_user']['token'],
+                reviewer=config['role_reviewer']['token'],
+            )
 
-        make_branch(repo_c, 'master', 'initial', {'f': 'c0'})
-        pr_c = make_pr(repo_c, 'C', [{'f': 'c1'}], label='a-thing')
+            make_branch(repo_c, 'master', 'initial', {'f': 'c0'})
+            pr_c = make_pr(
+                repo_c, 'a-thing', [{'f': 'c1'}],
+                user=config['role_user']['token'],
+                reviewer=config['role_reviewer']['token'],
+            )
+        env.run_crons()
 
-        run_crons(env)
         assert pr_a.comments == []
         assert pr_b.comments == [
             (users['reviewer'], 'hansen r+'),
@@ -337,26 +437,32 @@ class TestCompanionsNotReady:
              ))
         ]
 
-def test_other_failed(env, project, repo_a, repo_b, owner, users):
+def test_other_failed(env, project, repo_a, repo_b, users, config):
     """ In a non-matched-branch scenario, if the companion staging (copy of
     targets) fails when built with the PR, it should provide a non-useless
     message
     """
-    make_branch(repo_a, 'master', 'initial', {'a': 'a_0'})
-    # pr_a is born ready
-    pr_a = make_pr(repo_a, 'A', [{'a': 'a_1'}], label='do-a-thing')
+    with repo_a, repo_b:
+        make_branch(repo_a, 'master', 'initial', {'a': 'a_0'})
+        # pr_a is born ready
+        pr_a = make_pr(
+            repo_a, 'do-a-thing', [{'a': 'a_1'}],
+            user=config['role_user']['token'],
+            reviewer=config['role_reviewer']['token'],
+        )
 
-    make_branch(repo_b, 'master', 'initial', {'a': 'b_0'})
+        make_branch(repo_b, 'master', 'initial', {'a': 'b_0'})
+    env.run_crons()
 
-    run_crons(env)
     pr = to_pr(env, pr_a)
     assert pr.staging_id
 
-    repo_a.post_status('heads/staging.master', 'success', 'legal/cla')
-    repo_a.post_status('heads/staging.master', 'success', 'ci/runbot', target_url="http://example.org/a")
-    repo_b.post_status('heads/staging.master', 'success', 'legal/cla')
-    repo_b.post_status('heads/staging.master', 'failure', 'ci/runbot', target_url="http://example.org/b")
-    run_crons(env)
+    with repo_a, repo_b:
+        repo_a.post_status('heads/staging.master', 'success', 'legal/cla')
+        repo_a.post_status('heads/staging.master', 'success', 'ci/runbot', target_url="http://example.org/a")
+        repo_b.post_status('heads/staging.master', 'success', 'legal/cla')
+        repo_b.post_status('heads/staging.master', 'failure', 'ci/runbot', target_url="http://example.org/b")
+    env.run_crons()
 
     sth = repo_b.commit('heads/staging.master').id
     assert not pr.staging_id
@@ -367,22 +473,23 @@ def test_other_failed(env, project, repo_a, repo_b, owner, users):
     ]
 
 class TestMultiBatches:
-    def test_batching(self, env, project, repo_a, repo_b):
+    def test_batching(self, env, project, repo_a, repo_b, config):
         """ If multiple batches (label groups) are ready they should get batched
         together (within the limits of teh project's batch limit)
         """
         project.batch_limit = 3
-        make_branch(repo_a, 'master', 'initial', {'a': 'a0'})
-        make_branch(repo_b, 'master', 'initial', {'b': 'b0'})
 
-        prs = [(
-            a and to_pr(env, make_pr(repo_a, 'A{}'.format(i), [{'a{}'.format(i): 'a{}'.format(i)}], label='batch{}'.format(i))),
-            b and to_pr(env, make_pr(repo_b, 'B{}'.format(i), [{'b{}'.format(i): 'b{}'.format(i)}], label='batch{}'.format(i)))
-        )
-            for i, (a, b) in enumerate([(1, 1), (0, 1), (1, 1), (1, 1), (1, 0)])
-        ]
+        with repo_a, repo_b:
+            make_branch(repo_a, 'master', 'initial', {'a': 'a0'})
+            make_branch(repo_b, 'master', 'initial', {'b': 'b0'})
 
-        run_crons(env)
+            prs = [(
+                a and to_pr(env, make_pr(repo_a, 'batch{}'.format(i), [{'a{}'.format(i): 'a{}'.format(i)}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'],)),
+                b and to_pr(env, make_pr(repo_b, 'batch{}'.format(i), [{'b{}'.format(i): 'b{}'.format(i)}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'],))
+            )
+                for i, (a, b) in enumerate([(1, 1), (0, 1), (1, 1), (1, 1), (1, 0)])
+            ]
+        env.run_crons()
 
         st = env['runbot_merge.stagings'].search([])
         assert st
@@ -398,20 +505,20 @@ class TestMultiBatches:
         assert not prs[3][1].staging_id
         assert not prs[4][0].staging_id
 
-    def test_batching_split(self, env, repo_a, repo_b):
+    def test_batching_split(self, env, repo_a, repo_b, config):
         """ If a staging fails, it should get split properly across repos
         """
-        make_branch(repo_a, 'master', 'initial', {'a': 'a0'})
-        make_branch(repo_b, 'master', 'initial', {'b': 'b0'})
+        with repo_a, repo_b:
+            make_branch(repo_a, 'master', 'initial', {'a': 'a0'})
+            make_branch(repo_b, 'master', 'initial', {'b': 'b0'})
 
-        prs = [(
-            a and to_pr(env, make_pr(repo_a, 'A{}'.format(i), [{'a{}'.format(i): 'a{}'.format(i)}], label='batch{}'.format(i))),
-            b and to_pr(env, make_pr(repo_b, 'B{}'.format(i), [{'b{}'.format(i): 'b{}'.format(i)}], label='batch{}'.format(i)))
-        )
-            for i, (a, b) in enumerate([(1, 1), (0, 1), (1, 1), (1, 1), (1, 0)])
-        ]
-
-        run_crons(env)
+            prs = [(
+                a and to_pr(env, make_pr(repo_a, 'batch{}'.format(i), [{'a{}'.format(i): 'a{}'.format(i)}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'],)),
+                b and to_pr(env, make_pr(repo_b, 'batch{}'.format(i), [{'b{}'.format(i): 'b{}'.format(i)}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'],))
+            )
+                for i, (a, b) in enumerate([(1, 1), (0, 1), (1, 1), (1, 1), (1, 0)])
+            ]
+        env.run_crons()
 
         st0 = env['runbot_merge.stagings'].search([])
         assert len(st0.batch_ids) == 5
@@ -419,10 +526,10 @@ class TestMultiBatches:
 
         # mark b.staging as failed -> should create two splits with (0, 1)
         # and (2, 3, 4) and stage the first one
-        repo_b.post_status('heads/staging.master', 'success', 'legal/cla')
-        repo_b.post_status('heads/staging.master', 'failure', 'ci/runbot')
-
-        run_crons(env)
+        with repo_b:
+            repo_b.post_status('heads/staging.master', 'success', 'legal/cla')
+            repo_b.post_status('heads/staging.master', 'failure', 'ci/runbot')
+        env.run_crons()
 
         assert not st0.active
 
@@ -440,21 +547,21 @@ class TestMultiBatches:
         assert sp.mapped('batch_ids.prs') == \
             prs[2][0] | prs[2][1] | prs[3][0] | prs[3][1] | prs[4][0]
 
-def test_urgent(env, repo_a, repo_b):
+def test_urgent(env, repo_a, repo_b, config):
     """ Either PR of a co-dependent pair being p=0 leads to the entire pair
     being prioritized
     """
-    make_branch(repo_a, 'master', 'initial', {'a0': 'a'})
-    make_branch(repo_b, 'master', 'initial', {'b0': 'b'})
+    with repo_a, repo_b:
+        make_branch(repo_a, 'master', 'initial', {'a0': 'a'})
+        make_branch(repo_b, 'master', 'initial', {'b0': 'b'})
 
-    pr_a = make_pr(repo_a, 'A', [{'a1': 'a'}, {'a2': 'a'}], label='batch', reviewer=None, statuses=[])
-    pr_b = make_pr(repo_b, 'B', [{'b1': 'b'}, {'b2': 'b'}], label='batch', reviewer=None, statuses=[])
-    pr_c = make_pr(repo_a, 'C', [{'c1': 'c', 'c2': 'c'}])
+        pr_a = make_pr(repo_a, 'batch', [{'a1': 'a'}, {'a2': 'a'}], user=config['role_user']['token'], reviewer=None, statuses=[])
+        pr_b = make_pr(repo_b, 'batch', [{'b1': 'b'}, {'b2': 'b'}], user=config['role_user']['token'], reviewer=None, statuses=[])
+        pr_c = make_pr(repo_a, 'C', [{'c1': 'c', 'c2': 'c'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'],)
 
-    pr_a.post_comment('hansen rebase-merge', 'reviewer')
-    pr_b.post_comment('hansen rebase-merge p=0', 'reviewer')
-
-    run_crons(env)
+        pr_a.post_comment('hansen rebase-merge', config['role_reviewer']['token'])
+        pr_b.post_comment('hansen rebase-merge p=0', config['role_reviewer']['token'])
+    env.run_crons()
     # should have batched pr_a and pr_b despite neither being reviewed or
     # approved
     p_a, p_b = to_pr(env, pr_a), to_pr(env, pr_b)
@@ -464,79 +571,81 @@ def test_urgent(env, repo_a, repo_b):
     assert not p_c.staging_id
 
 class TestBlocked:
-    def test_merge_method(self, env, repo_a):
-        make_branch(repo_a, 'master', 'initial', {'a0': 'a'})
+    def test_merge_method(self, env, repo_a, config):
+        with repo_a:
+            make_branch(repo_a, 'master', 'initial', {'a0': 'a'})
 
-        pr = make_pr(repo_a, 'A', [{'a1': 'a'}, {'a2': 'a'}])
-
-        run_crons(env)
+            pr = make_pr(repo_a, 'A', [{'a1': 'a'}, {'a2': 'a'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'],)
+        env.run_crons()
 
         p = to_pr(env, pr)
         assert p.state == 'ready'
-        print(p.id, p.squash, p.merge_method)
         assert p.blocked
 
-        pr.post_comment('hansen rebase-merge', 'reviewer')
+        with repo_a: pr.post_comment('hansen rebase-merge', config['role_reviewer']['token'])
         assert not p.blocked
 
-    def test_linked_closed(self, env, repo_a, repo_b):
-        make_branch(repo_a, 'master', 'initial', {'a0': 'a'})
-        make_branch(repo_b, 'master', 'initial', {'b0': 'b'})
+    def test_linked_closed(self, env, repo_a, repo_b, config):
+        with repo_a, repo_b:
+            make_branch(repo_a, 'master', 'initial', {'a0': 'a'})
+            make_branch(repo_b, 'master', 'initial', {'b0': 'b'})
 
-        pr = make_pr(repo_a, 'A', [{'a1': 'a'}], label='xxx')
-        b = make_pr(repo_b, 'B', [{'b1': 'b'}], label='xxx', statuses=[])
-        run_crons(env)
+            pr = make_pr(repo_a, 'xxx', [{'a1': 'a'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'],)
+            b = make_pr(repo_b, 'xxx', [{'b1': 'b'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'], statuses=[])
+        env.run_crons()
 
         p = to_pr(env, pr)
         assert p.blocked
-        b.close()
+        with repo_b: b.close()
         # FIXME: find a way for PR.blocked to depend on linked PR somehow so this isn't needed
         p.invalidate_cache(['blocked'], [p.id])
         assert not p.blocked
 
-    def test_linked_merged(self, env, repo_a, repo_b):
-        make_branch(repo_a, 'master', 'initial', {'a0': 'a'})
-        make_branch(repo_b, 'master', 'initial', {'b0': 'b'})
+    def test_linked_merged(self, env, repo_a, repo_b, config):
+        with repo_a, repo_b:
+            make_branch(repo_a, 'master', 'initial', {'a0': 'a'})
+            make_branch(repo_b, 'master', 'initial', {'b0': 'b'})
 
-        b = make_pr(repo_b, 'B', [{'b1': 'b'}], label='xxx')
+            b = make_pr(repo_b, 'xxx', [{'b1': 'b'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'],)
+        env.run_crons() # stage b and c
 
-        run_crons(env) # stage b and c
-
-        repo_a.post_status('heads/staging.master', 'success', 'legal/cla')
-        repo_a.post_status('heads/staging.master', 'success', 'ci/runbot')
-        repo_b.post_status('heads/staging.master', 'success', 'legal/cla')
-        repo_b.post_status('heads/staging.master', 'success', 'ci/runbot')
-
-        run_crons(env) # merge b and c
+        with repo_a, repo_b:
+            repo_a.post_status('heads/staging.master', 'success', 'legal/cla')
+            repo_a.post_status('heads/staging.master', 'success', 'ci/runbot')
+            repo_b.post_status('heads/staging.master', 'success', 'legal/cla')
+            repo_b.post_status('heads/staging.master', 'success', 'ci/runbot')
+        env.run_crons() # merge b and c
         assert to_pr(env, b).state == 'merged'
 
-        pr = make_pr(repo_a, 'A', [{'a1': 'a'}], label='xxx')
-        run_crons(env) # merge b and c
+        with repo_a:
+            pr = make_pr(repo_a, 'xxx', [{'a1': 'a'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'],)
+        env.run_crons() # merge b and c
 
         p = to_pr(env, pr)
         assert not p.blocked
 
-    def test_linked_unready(self, env, repo_a, repo_b):
+    def test_linked_unready(self, env, repo_a, repo_b, config):
         """ Create a PR A linked to a non-ready PR B,
         * A is blocked by default
         * A is not blocked if A.p=0
         * A is not blocked if B.p=0
         """
-        make_branch(repo_a, 'master', 'initial', {'a0': 'a'})
-        make_branch(repo_b, 'master', 'initial', {'b0': 'b'})
+        with repo_a, repo_b:
+            make_branch(repo_a, 'master', 'initial', {'a0': 'a'})
+            make_branch(repo_b, 'master', 'initial', {'b0': 'b'})
 
-        a = make_pr(repo_a, 'A', [{'a1': 'a'}], label='xxx')
-        b = make_pr(repo_b, 'B', [{'b1': 'b'}], label='xxx', statuses=[])
-        run_crons(env)
+            a = make_pr(repo_a, 'xxx', [{'a1': 'a'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'],)
+            b = make_pr(repo_b, 'xxx', [{'b1': 'b'}], user=config['role_user']['token'], reviewer=config['role_reviewer']['token'], statuses=[])
+        env.run_crons()
 
         pr_a = to_pr(env, a)
         assert pr_a.blocked
 
-        a.post_comment('hansen p=0', 'reviewer')
+        with repo_a: a.post_comment('hansen p=0', config['role_reviewer']['token'])
         assert not pr_a.blocked
 
-        a.post_comment('hansen p=2', 'reviewer')
+        with repo_a: a.post_comment('hansen p=2', config['role_reviewer']['token'])
         assert pr_a.blocked
 
-        b.post_comment('hansen p=0', 'reviewer')
+        with repo_b: b.post_comment('hansen p=0', config['role_reviewer']['token'])
         assert not pr_a.blocked
